Check token expiration time

Checks whether the token will expire soon or has already expired and throws a warning to the user.

Usage

check_token_expiration_time(token, auto_update = TRUE)

Arguments

token

(list) Access token to the API obtained from get_token()

auto_update

Attempts to update token automatically

Value

NULL It checks whether the token has expired or is soon to expire and alerts the user.