Verify and Convert Place of Occurrence External Cause Code

This function takes one parameter, which can be a number or a string. If it is a number, the function verifies it against predefined codes. If it is not one of the codes, it returns an error. If it is a text, the function returns the number corresponding to the code. If it is NULL, the function returns NULL.

Usage

code_place(input)

Arguments

input

Either a number or a string representing the place of occurrence external cause code.

Value

The corresponding place of occurrence external cause code as a number, or an error if the input is invalid.