Search info in a tidy way

Searches for chapter, block or code in a vectorized fashion and returns into a tidy object (as a mutate would)

.icd10_search_tidy(
  .data,
  colname,
  searchfun,
  token,
  release,
  language,
  dry_run,
  codes_only,
  ...
)

Arguments

.data

A data.frame or a data.frameextension such as a tibble.

colname

(character) Name of the data.frame column with the ICD-10 code/block or chapter to search

searchfun

Search function to be utilized

token

(list) Access token to the API obtained from get_token()

release

(character) ICD-10 release (default = 2019). Use icd10_releases() to get all available releases.

language

(character) Language for the request. Available languages are listed in https://icd.who.int/docs/icd-api/SupportedClassifications/ the default is English (language = "en").

dry_run

(logical) If TRUE will create the request and generate an empty token. This function is mostly for development.

codes_only

Return only the codes without the titles. This option is faster as it realizes less requests.

...

Additional parameters to pass to searchfun

Value

A data frame containing all the parent nodes for the block/code/chapter as well as the original dataframe