Chapters of ICD-10

Lists the available chapters for an ICD-10 release.

Usage

icd10_chapters(token, release = 2019, language = "en", codes_only = FALSE)

Arguments

token

(list) Access token to the API obtained from get_token()

release

(character) ICD-10 release (default = 2019). Use icd10_releases() to get all available releases.

language

(character) Language for the request. Available languages are listed in https://icd.who.int/docs/icd-api/SupportedClassifications/ the default is English (language = "en").

codes_only

Return only the codes without the titles. This option is faster as it realizes less requests.

Value

Data frame with all the chapters and their description

Examples

#Assuming that the CLIENT ID and CLIENT SECRET are set up. Substitute accordingly
if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {
  #Generated token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)

  # Get title and url for release
  icd10_chapters(token, release = 2016)
}
#>    codes
#> 1      I
#> 2     II
#> 3    III
#> 4     IV
#> 5      V
#> 6     VI
#> 7    VII
#> 8   VIII
#> 9     IX
#> 10     X
#> 11    XI
#> 12   XII
#> 13  XIII
#> 14   XIV
#> 15    XV
#> 16   XVI
#> 17  XVII
#> 18 XVIII
#> 19   XIX
#> 20    XX
#> 21   XXI
#> 22  XXII
#>                                                                                                  title
#> 1                                                            Certain infectious and parasitic diseases
#> 2                                                                                            Neoplasms
#> 3  Diseases of the blood and blood-forming organs and certain disorders involving the immune mechanism
#> 4                                                        Endocrine, nutritional and metabolic diseases
#> 5                                                                     Mental and behavioural disorders
#> 6                                                                       Diseases of the nervous system
#> 7                                                                       Diseases of the eye and adnexa
#> 8                                                              Diseases of the ear and mastoid process
#> 9                                                                   Diseases of the circulatory system
#> 10                                                                  Diseases of the respiratory system
#> 11                                                                    Diseases of the digestive system
#> 12                                                        Diseases of the skin and subcutaneous tissue
#> 13                                        Diseases of the musculoskeletal system and connective tissue
#> 14                                                                Diseases of the genitourinary system
#> 15                                                            Pregnancy, childbirth and the puerperium
#> 16                                              Certain conditions originating in the perinatal period
#> 17                                Congenital malformations, deformations and chromosomal abnormalities
#> 18             Symptoms, signs and abnormal clinical and laboratory findings, not elsewhere classified
#> 19                                 Injury, poisoning and certain other consequences of external causes
#> 20                                                          External causes of morbidity and mortality
#> 21                                  Factors influencing health status and contact with health services
#> 22                                                                          Codes for special purposes