Search the ICD-10 for information recursively

Search the ICD-10 using the WHO API for information on blocks, chapters or codes.

icd10_search(
  token,
  search_vals,
  release = 2019,
  language = "en",
  auto_update = TRUE
)

Arguments

token

(list) Access token to the API obtained from get_token()

search_vals

Either codes, blocks or chapter values to search.

release

(character) ICD-10 release (default = 2019). Use icd10_releases() to get all available releases.

language

(character) Language for the request. Available languages are listed in https://icd.who.int/docs/icd-api/SupportedClassifications/ the default is English (language = "en").

auto_update

Attempts to update token automatically

Details

Searches for either codes, blocks or chapters in a specified release.

Examples

#Assuming that the CLIENT ID and CLIENT SECRET are set up. Substitute accordingly
if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {
  #Generated token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)

  # Search multiple titles
  dbf <- icd10_search(token, c("D50.1","I14","XII","D50","D60","HEY", "D50-D53", NA_character_))
  dbf <- icd10_search(token, c("XI","XII", NA_character_))
  dbf <- icd10_search(token, c("D50-D53", "D50-D53", NA_character_, "D60-D64"))

}
#> Warning: Request not found. Possibly any of release, chapter/block/code or language is not available or incorrectly specified.
#> Warning: Request not found. Possibly any of release, chapter/block/code or language is not available or incorrectly specified.
#> Warning: Request not found. Possibly any of release, chapter/block/code or language is not available or incorrectly specified.
#> Warning: Request not found. Possibly any of release, chapter/block/code or language is not available or incorrectly specified.