Identify an URI in ICD-11

Provides all of the characteristics from an URI

icd11_id(
  token,
  id,
  release = "2024-01",
  include = NULL,
  language = "en",
  auto_update = TRUE
)

Arguments

token

(list) Access token to the API obtained from get_token()

id

(stirng) String either with a complete URI of the last digits of the URI

release

Release version as listed in https://icd.who.int/docs/icd-api/SupportedClassifications/#supported-classifications-and-versions the default is 2024-01.

include

Either NULL of provide additional information such as ancestor, decendant or diagnosticCriteria. For example, if you set include="ancestor" the returned response will include all ancestor entity URIs for the requested entity. Use a vector to include multiple properties. For example include = c("ancestor", "diagnosticCriteria")

language

(character) Language for the request. Available languages are listed in https://icd.who.int/docs/icd-api/SupportedClassifications/ the default is English (language = "en").

auto_update

Attempts to update token automatically

Value

Character vector with the ICD-11 URI and its characteristics

Examples

# Assuming that the CLIENT ID and CLIENT SECRET are set up. Substitute accordingly
if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {
  # Generated token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)

  # Search for ICD-11 URI
  icd11_id(token, id = "http://id.who.int/icd/entity/1442995018")

  # Search for ICD-11 URI with parents and diagnostic criteria
  icd11_id(token,
    id = "http://id.who.int/icd/entity/455013390",
    include = c("ancestor", "diagnosticCriteria")
  )
}
#> $`@context`
#> [1] "http://id.who.int/icd/contexts/contextForFoundationEntity.json"
#> 
#> $`@id`
#> [1] "http://id.who.int/icd/entity/455013390"
#> 
#> $parent
#> $parent[[1]]
#> [1] "http://id.who.int/icd/entity/448895267"
#> 
#> 
#> $child
#> $child[[1]]
#> [1] "http://id.who.int/icd/entity/1435254666"
#> 
#> $child[[2]]
#> [1] "http://id.who.int/icd/entity/1630407678"
#> 
#> $child[[3]]
#> [1] "http://id.who.int/icd/entity/1766440644"
#> 
#> $child[[4]]
#> [1] "http://id.who.int/icd/entity/1954798891"
#> 
#> $child[[5]]
#> [1] "http://id.who.int/icd/entity/21500692"
#> 
#> $child[[6]]
#> [1] "http://id.who.int/icd/entity/334423054"
#> 
#> $child[[7]]
#> [1] "http://id.who.int/icd/entity/274880002"
#> 
#> $child[[8]]
#> [1] "http://id.who.int/icd/entity/1296093776"
#> 
#> $child[[9]]
#> [1] "http://id.who.int/icd/entity/868865918"
#> 
#> $child[[10]]
#> [1] "http://id.who.int/icd/entity/1218729044"
#> 
#> $child[[11]]
#> [1] "http://id.who.int/icd/entity/426429380"
#> 
#> $child[[12]]
#> [1] "http://id.who.int/icd/entity/197934298"
#> 
#> $child[[13]]
#> [1] "http://id.who.int/icd/entity/1256772020"
#> 
#> $child[[14]]
#> [1] "http://id.who.int/icd/entity/1639304259"
#> 
#> $child[[15]]
#> [1] "http://id.who.int/icd/entity/1473673350"
#> 
#> $child[[16]]
#> [1] "http://id.who.int/icd/entity/30659757"
#> 
#> $child[[17]]
#> [1] "http://id.who.int/icd/entity/577470983"
#> 
#> $child[[18]]
#> [1] "http://id.who.int/icd/entity/714000734"
#> 
#> $child[[19]]
#> [1] "http://id.who.int/icd/entity/1306203631"
#> 
#> $child[[20]]
#> [1] "http://id.who.int/icd/entity/223744320"
#> 
#> $child[[21]]
#> [1] "http://id.who.int/icd/entity/1843895818"
#> 
#> $child[[22]]
#> [1] "http://id.who.int/icd/entity/435227771"
#> 
#> $child[[23]]
#> [1] "http://id.who.int/icd/entity/850137482"
#> 
#> $child[[24]]
#> [1] "http://id.who.int/icd/entity/1249056269"
#> 
#> $child[[25]]
#> [1] "http://id.who.int/icd/entity/1596590595"
#> 
#> $child[[26]]
#> [1] "http://id.who.int/icd/entity/718687701"
#> 
#> $child[[27]]
#> [1] "http://id.who.int/icd/entity/231358748"
#> 
#> $child[[28]]
#> [1] "http://id.who.int/icd/entity/979408586"
#> 
#> 
#> $browserUrl
#> [1] "https://icd.who.int/browse/2024-01/foundation/en#455013390"
#> 
#> $ancestor
#> $ancestor[[1]]
#> [1] "http://id.who.int/icd/entity/448895267"
#> 
#> 
#> $title
#> $title$`@language`
#> [1] "en"
#> 
#> $title$`@value`
#> [1] "ICD Category"
#> 
#>