Create and perform a request to the WHO API v2.

This is the main function for development. Creates a request for the API using make_request and then performs the request using httr2::req_perform. Finally it obtains the body of the response and returns to user.

Usage

request_WHO(
  url,
  token = NULL,
  language = "en",
  auto_update = TRUE,
  req_body = NULL,
  warning_message_404 = "url for request not found",
  method = "GET",
  post_process_function = function(x) {
     x
 }
)

Arguments

url

(character) Address for the request

token

(list) Access token to the API obtained from get_token()

language

(character) Language for the request. Available languages are listed in https://icd.who.int/docs/icd-api/SupportedClassifications/ the default is English (language = "en").

auto_update

Attempts to update token automatically

post_process_function

Function to apply to the body of the response after obtaining a response.

Value

The response's body from the request

Examples

if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {

  #Get the URL
  url <- "https://id.who.int/icd/release/10"

  #Generated token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)

  # Get title and url for release
  request_WHO(url, token = token)
}
#> $`@context`
#> [1] "http://id.who.int/icd/contexts/contextForMultiVersion.json"
#> 
#> $`@id`
#> [1] "http://id.who.int/icd/release/10"
#> 
#> $title
#> $title$`@language`
#> [1] "en"
#> 
#> $title$`@value`
#> [1] "International Statistical Classification of Diseases and Related Health Problems 10th Revision (ICD-10)"
#> 
#> 
#> $latestRelease
#> [1] "http://id.who.int/icd/release/10/2019"
#> 
#> $release
#> $release[[1]]
#> [1] "http://id.who.int/icd/release/10/2019"
#> 
#> $release[[2]]
#> [1] "http://id.who.int/icd/release/10/2016"
#> 
#> $release[[3]]
#> [1] "http://id.who.int/icd/release/10/2010"
#> 
#> $release[[4]]
#> [1] "http://id.who.int/icd/release/10/2008"
#> 
#>