Search for information on multiple codes, blocks or chapter in a data.frame or tibble.

Search for information on multiple codes, blocks or chapter in a data.frame or tibble.

icd10_code_info_tidy(
  .data,
  colname,
  token,
  release = 2019,
  language = "en",
  dry_run = FALSE,
  codes_only = FALSE,
  validate_code = TRUE
)

icd10_block_info_tidy(
  .data,
  colname,
  token,
  release = 2019,
  language = "en",
  dry_run = FALSE,
  codes_only = FALSE,
  validate_code = TRUE
)

icd10_chapter_info_tidy(
  .data,
  colname,
  token,
  release = 2019,
  language = "en",
  dry_run = FALSE,
  codes_only = FALSE,
  validate_code = TRUE
)

Arguments

.data

A data.frame or a data.frameextension such as a tibble.

colname

(character) Name of the data.frame column with the ICD-10 code/block or chapter to search

token

(list) Access token to the API obtained from get_token()

release

(character) ICD-10 release (default = 2019). Use icd10_releases() to get all available releases.

language

(character) Language for the request. Available languages are listed in https://icd.who.int/docs/icd-api/SupportedClassifications/ the default is English (language = "en").

dry_run

(logical) If TRUE will create the request and generate an empty token. This function is mostly for development.

codes_only

Return only the codes without the titles. This option is faster as it realizes less requests.

validate_code

( logical) Apply additional functions to validate that the ICD-10 code is written as the API needs it.

Examples

#Change `dry_run = FALSE` to run the example
token <- get_token("123","123", dry_run = TRUE)
#> POST /connect/token HTTP/1.1
#> Host: icdaccessmanagement.who.int
#> User-Agent: httr2/1.0.1 r-curl/5.2.1 libcurl/8.4.0
#> Accept: */*
#> Accept-Encoding: deflate, gzip
#> Content-Type: application/x-www-form-urlencoded
#> Content-Length: 81
#> 
#> client_id=123&client_secret=123&scope=icdapi_access&grant_type=client_credentials
codes_df <- data.frame(Sex = c("M","F","F"), icd10 = c("E14.1","C80.0","F14"))
codes_df |>
    icd10_code_info_tidy("icd10", token, dry_run = TRUE)
#> [1] "Searching code: E14.1"
#> GET /icd/release/10/2019/E14.1 HTTP/1.1
#> Host: id.who.int
#> User-Agent: httr2/1.0.1 r-curl/5.2.1 libcurl/8.4.0
#> Accept-Encoding: deflate, gzip
#> Accept: application/json
#> Accept-Language: en
#> API-Version: v2
#> 
#> GET /icd/release/10/2019/E14.1 HTTP/1.1
#> Host: id.who.int
#> User-Agent: httr2/1.0.1 r-curl/5.2.1 libcurl/8.4.0
#> Accept-Encoding: deflate, gzip
#> Accept: application/json
#> Accept-Language: en
#> API-Version: v2
#> 
#> [1] "Searching code: C80.0"
#> GET /icd/release/10/2019/C80.0 HTTP/1.1
#> Host: id.who.int
#> User-Agent: httr2/1.0.1 r-curl/5.2.1 libcurl/8.4.0
#> Accept-Encoding: deflate, gzip
#> Accept: application/json
#> Accept-Language: en
#> API-Version: v2
#> 
#> GET /icd/release/10/2019/C80.0 HTTP/1.1
#> Host: id.who.int
#> User-Agent: httr2/1.0.1 r-curl/5.2.1 libcurl/8.4.0
#> Accept-Encoding: deflate, gzip
#> Accept: application/json
#> Accept-Language: en
#> API-Version: v2
#> 
#> [1] "Searching code: F14"
#> GET /icd/release/10/2019/F14 HTTP/1.1
#> Host: id.who.int
#> User-Agent: httr2/1.0.1 r-curl/5.2.1 libcurl/8.4.0
#> Accept-Encoding: deflate, gzip
#> Accept: application/json
#> Accept-Language: en
#> API-Version: v2
#> 
#> GET /icd/release/10/2019/F14 HTTP/1.1
#> Host: id.who.int
#> User-Agent: httr2/1.0.1 r-curl/5.2.1 libcurl/8.4.0
#> Accept-Encoding: deflate, gzip
#> Accept: application/json
#> Accept-Language: en
#> API-Version: v2
#> 
#> Warning: No value of `searchvec` was found
#>   Sex icd10
#> 1   M E14.1
#> 2   F C80.0
#> 3   F   F14
#Change `dry_run = FALSE` to run the example
token <- get_token("123","123", dry_run = TRUE)
#> POST /connect/token HTTP/1.1
#> Host: icdaccessmanagement.who.int
#> User-Agent: httr2/1.0.1 r-curl/5.2.1 libcurl/8.4.0
#> Accept: */*
#> Accept-Encoding: deflate, gzip
#> Content-Type: application/x-www-form-urlencoded
#> Content-Length: 81
#> 
#> client_id=123&client_secret=123&scope=icdapi_access&grant_type=client_credentials
codes_df <- data.frame(Sex = c("M","F","F"), icd10 = c("C76-C80","E10-E14","F10-F19"))
codes_df |>
    icd10_block_info_tidy("icd10", token, dry_run = TRUE)
#> [1] "Searching code: C76-C80"
#> GET /icd/release/10/2019/C76-C80 HTTP/1.1
#> Host: id.who.int
#> User-Agent: httr2/1.0.1 r-curl/5.2.1 libcurl/8.4.0
#> Accept-Encoding: deflate, gzip
#> Accept: application/json
#> Accept-Language: en
#> API-Version: v2
#> 
#> GET /icd/release/10/2019/C76-C80 HTTP/1.1
#> Host: id.who.int
#> User-Agent: httr2/1.0.1 r-curl/5.2.1 libcurl/8.4.0
#> Accept-Encoding: deflate, gzip
#> Accept: application/json
#> Accept-Language: en
#> API-Version: v2
#> 
#> [1] "Searching code: E10-E14"
#> GET /icd/release/10/2019/E10-E14 HTTP/1.1
#> Host: id.who.int
#> User-Agent: httr2/1.0.1 r-curl/5.2.1 libcurl/8.4.0
#> Accept-Encoding: deflate, gzip
#> Accept: application/json
#> Accept-Language: en
#> API-Version: v2
#> 
#> GET /icd/release/10/2019/E10-E14 HTTP/1.1
#> Host: id.who.int
#> User-Agent: httr2/1.0.1 r-curl/5.2.1 libcurl/8.4.0
#> Accept-Encoding: deflate, gzip
#> Accept: application/json
#> Accept-Language: en
#> API-Version: v2
#> 
#> [1] "Searching code: F10-F19"
#> GET /icd/release/10/2019/F10-F19 HTTP/1.1
#> Host: id.who.int
#> User-Agent: httr2/1.0.1 r-curl/5.2.1 libcurl/8.4.0
#> Accept-Encoding: deflate, gzip
#> Accept: application/json
#> Accept-Language: en
#> API-Version: v2
#> 
#> GET /icd/release/10/2019/F10-F19 HTTP/1.1
#> Host: id.who.int
#> User-Agent: httr2/1.0.1 r-curl/5.2.1 libcurl/8.4.0
#> Accept-Encoding: deflate, gzip
#> Accept: application/json
#> Accept-Language: en
#> API-Version: v2
#> 
#> Warning: No value of `searchvec` was found
#>   Sex   icd10
#> 1   M C76-C80
#> 2   F E10-E14
#> 3   F F10-F19
#Change `dry_run = FALSE` to run the example
token    <- get_token("123","123", dry_run = TRUE)
#> POST /connect/token HTTP/1.1
#> Host: icdaccessmanagement.who.int
#> User-Agent: httr2/1.0.1 r-curl/5.2.1 libcurl/8.4.0
#> Accept: */*
#> Accept-Encoding: deflate, gzip
#> Content-Type: application/x-www-form-urlencoded
#> Content-Length: 81
#> 
#> client_id=123&client_secret=123&scope=icdapi_access&grant_type=client_credentials
codes_df <- data.frame(Sex = c("M","F","F"), icd10_chapter = c("X", "IV","II"))
codes_df |>
    icd10_chapter_info_tidy("icd10_chapter", token, dry_run = TRUE)
#> [1] "Searching code: X"
#> GET /icd/release/10/2019/X HTTP/1.1
#> Host: id.who.int
#> User-Agent: httr2/1.0.1 r-curl/5.2.1 libcurl/8.4.0
#> Accept-Encoding: deflate, gzip
#> Accept: application/json
#> Accept-Language: en
#> API-Version: v2
#> 
#> [1] "Searching code: IV"
#> GET /icd/release/10/2019/IV HTTP/1.1
#> Host: id.who.int
#> User-Agent: httr2/1.0.1 r-curl/5.2.1 libcurl/8.4.0
#> Accept-Encoding: deflate, gzip
#> Accept: application/json
#> Accept-Language: en
#> API-Version: v2
#> 
#> [1] "Searching code: II"
#> GET /icd/release/10/2019/II HTTP/1.1
#> Host: id.who.int
#> User-Agent: httr2/1.0.1 r-curl/5.2.1 libcurl/8.4.0
#> Accept-Encoding: deflate, gzip
#> Accept: application/json
#> Accept-Language: en
#> API-Version: v2
#> 
#> Warning: No value of `searchvec` was found
#>   Sex icd10_chapter
#> 1   M             X
#> 2   F            IV
#> 3   F            II