Block of ICD-10 Chapter

Lists the available sections inside a chapter for an ICD-10 release.

icd10_blocks(
  token,
  chapter,
  release = 2019,
  language = "en",
  dry_run = FALSE,
  codes_only = FALSE
)

Arguments

token

(list) Access token to the API obtained from get_token()

chapter

(character) ICD-10 chapter as a roman numeral

release

(character) ICD-10 release (default = 2019). Use icd10_releases() to get all available releases.

language

(character) Language for the request. Available languages are listed in https://icd.who.int/docs/icd-api/SupportedClassifications/ the default is English (language = "en").

dry_run

(logical) If TRUE will create the request and generate an empty token. This function is mostly for development.

codes_only

Return only the codes without the titles. This option is faster as it realizes less requests.

Value

Data frame with all the blocks inside a chapter and their description

Examples

#Change `dry_run = FALSE` to run the example
token <- get_token("123","123", dry_run = TRUE)
#> POST /connect/token HTTP/1.1
#> Host: icdaccessmanagement.who.int
#> User-Agent: httr2/1.0.1 r-curl/5.2.1 libcurl/8.4.0
#> Accept: */*
#> Accept-Encoding: deflate, gzip
#> Content-Type: application/x-www-form-urlencoded
#> Content-Length: 81
#> 
#> client_id=123&client_secret=123&scope=icdapi_access&grant_type=client_credentials
icd10_blocks(token, chapter = "IX", release = 2016, dry_run = TRUE)
#> GET /icd/release/10/2016/IX HTTP/1.1
#> Host: id.who.int
#> User-Agent: httr2/1.0.1 r-curl/5.2.1 libcurl/8.4.0
#> Accept-Encoding: deflate, gzip
#> Accept: application/json
#> Accept-Language: en
#> API-Version: v2
#> 
#> NULL