ICD-10 parent search

Generates a request to https://id.who.int/icd/release/10/\{release\}/\{site\}, and obtains the parent node.

icd10_parent(
  token,
  site,
  release = 2019,
  language = "en",
  codes_only = FALSE,
  auto_update = TRUE
)

Arguments

token

(list) Access token to the API obtained from get_token()

site

Specific link according to the WHO Swagger API https://id.who.int/swagger/index.html

release

(character) ICD-10 release (default = 2019). Use icd10_releases() to get all available releases.

language

(character) Language for the request. Available languages are listed in https://icd.who.int/docs/icd-api/SupportedClassifications/ the default is English (language = "en").

codes_only

Return only the codes without the titles. This option is faster as it realizes less requests.

Value

A character containing the site of the parent node.