Identify an URI in ICD-11

Provides all of the characteristics from an URI

icd11_id(
  token,
  id,
  linearization = "mms",
  release = "2024-01",
  include = NULL,
  language = "en",
  auto_update = TRUE
)

icd11_id2(
  token,
  id,
  release = "2024-01",
  include = NULL,
  language = "en",
  auto_update = TRUE
)

icd11_id3(
  token,
  id,
  linearization = "mms",
  language = "en",
  auto_update = TRUE
)

Arguments

token

(list) Access token to the API obtained from get_token()

id

(stirng) String either with a complete URI of the last digits of the URI

linearization

(character) ICD-11 linearization (default = "mms"). See all linearizations in the WHO's API website https://icd.who.int/docs/icd-api/APIDoc-Version2/#the-icd11-linearizations

release

Release version as listed in https://icd.who.int/docs/icd-api/SupportedClassifications/#supported-classifications-and-versions the default is 2024-01.

include

Either NULL of provide additional information such as ancestor, decendant or diagnosticCriteria. For example, if you set include="ancestor" the returned response will include all ancestor entity URIs for the requested entity. Use a vector to include multiple properties. For example include = c("ancestor", "diagnosticCriteria")

language

(character) Language for the request. Available languages are listed in https://icd.who.int/docs/icd-api/SupportedClassifications/ the default is English (language = "en").

auto_update

Attempts to update token automatically

Value

Character vector with the ICD-11 URI and its characteristics

Details

icd_id() Searches for a specific URI across a specific linearization an release. icd_id2() searches for a URI across a specific release regardless of linearization. Finally, icd_id3 searches for the URI across all ICD-11 releases.

Examples

# Assuming that the CLIENT ID and CLIENT SECRET are set up. Substitute accordingly
if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {
  # Generated token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)

  # Search for ICD-11 URI
  icd11_id(token, id = "http://id.who.int/icd/entity/1442995018")

  # Search for ICD-11 URI with parents and diagnostic criteria
  icd11_id(token,
    id = "http://id.who.int/icd/entity/1442995018",
    include = c("ancestor", "diagnosticCriteria")
  )
}
#> $`@context`
#> [1] "http://id.who.int/icd/contexts/contextForLinearizationEntity.json"
#> 
#> $`@id`
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1442995018"
#> 
#> $parent
#> $parent[[1]]
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/843843448"
#> 
#> 
#> $browserUrl
#> [1] "https://icd.who.int/browse/2024-01/mms/en#1442995018"
#> 
#> $code
#> [1] "8B20"
#> 
#> $source
#> [1] "http://id.who.int/icd/entity/1442995018"
#> 
#> $classKind
#> [1] "category"
#> 
#> $postcoordinationScale
#> $postcoordinationScale[[1]]
#> $postcoordinationScale[[1]]$`@id`
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1442995018/postcoordinationScale/laterality"
#> 
#> $postcoordinationScale[[1]]$axisName
#> [1] "http://id.who.int/icd/schema/laterality"
#> 
#> $postcoordinationScale[[1]]$requiredPostcoordination
#> [1] "false"
#> 
#> $postcoordinationScale[[1]]$allowMultipleValues
#> [1] "NotAllowed"
#> 
#> $postcoordinationScale[[1]]$scaleEntity
#> $postcoordinationScale[[1]]$scaleEntity[[1]]
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/627678743"
#> 
#> $postcoordinationScale[[1]]$scaleEntity[[2]]
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/271422288"
#> 
#> $postcoordinationScale[[1]]$scaleEntity[[3]]
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/876572005"
#> 
#> $postcoordinationScale[[1]]$scaleEntity[[4]]
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1038788978"
#> 
#> 
#> 
#> $postcoordinationScale[[2]]
#> $postcoordinationScale[[2]]$`@id`
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1442995018/postcoordinationScale/hasManifestation"
#> 
#> $postcoordinationScale[[2]]$axisName
#> [1] "http://id.who.int/icd/schema/hasManifestation"
#> 
#> $postcoordinationScale[[2]]$requiredPostcoordination
#> [1] "false"
#> 
#> $postcoordinationScale[[2]]$allowMultipleValues
#> [1] "AllowAlways"
#> 
#> $postcoordinationScale[[2]]$scaleEntity
#> $postcoordinationScale[[2]]$scaleEntity[[1]]
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1173205716"
#> 
#> 
#> 
#> 
#> $ancestor
#> $ancestor[[1]]
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/843843448"
#> 
#> $ancestor[[2]]
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1296093776"
#> 
#> 
#> $diagnosticCriteria
#> $diagnosticCriteria$`@language`
#> [1] "en"
#> 
#> $diagnosticCriteria$`@value`
#> [1] "Signs and symtoms:&nbsp;\n\n\n\n\n\n\n<!--[if gte mso 9]><xml>\n <o:DocumentProperties>\n <o:Revision>0</o:Revision>\n <o:TotalTime>0</o:TotalTime>\n <o:Pages>1</o:Pages>\n <o:Words>16</o:Words>\n <o:Characters>92</o:Characters>\n <o:Company>University of California, Berkeley</o:Company>\n <o:Lines>1</o:Lines>\n <o:Paragraphs>1</o:Paragraphs>\n <o:CharactersWithSpaces>107</o:CharactersWithSpaces>\n <o:Version>14.0</o:Version>\n </o:DocumentProperties>\n</xml><![endif]-->\n\n<!--[if gte mso 9]><xml>\n <w:WordDocument>\n <w:View>Normal</w:View>\n <w:Zoom>0</w:Zoom>\n <w:TrackMoves/>\n <w:TrackFormatting/>\n <w:PunctuationKerning/>\n <w:ValidateAgainstSchemas/>\n <w:SaveIfXMLInvalid>false</w:SaveIfXMLInvalid>\n <w:IgnoreMixedContent>false</w:IgnoreMixedContent>\n <w:AlwaysShowPlaceholderText>false</w:AlwaysShowPlaceholderText>\n <w:DoNotPromoteQF/>\n <w:LidThemeOther>EN-US</w:LidThemeOther>\n <w:LidThemeAsian>JA</w:LidThemeAsian>\n <w:LidThemeComplexScript>X-NONE</w:LidThemeComplexScript>\n <w:Compatibility>\n <w:BreakWrappedTables/>\n <w:SnapToGridInCell/>\n <w:WrapTextWithPunct/>\n <w:UseAsianBreakRules/>\n <w:DontGrowAutofit/>\n <w:SplitPgBreakAndParaMark/>\n <w:EnableOpenTypeKerning/>\n <w:DontFlipMirrorIndents/>\n <w:OverrideTableStyleHps/>\n </w:Compatibility>\n <m:mathPr>\n <m:mathFont m:val=\"Cambria Math\"/>\n <m:brkBin m:val=\"before\"/>\n <m:brkBinSub m:val=\"&#45;-\"/>\n <m:smallFrac m:val=\"off\"/>\n <m:dispDef/>\n <m:lMargin m:val=\"0\"/>\n <m:rMargin m:val=\"0\"/>\n <m:defJc m:val=\"centerGroup\"/>\n <m:wrapIndent m:val=\"1440\"/>\n <m:intLim m:val=\"subSup\"/>\n <m:naryLim m:val=\"undOvr\"/>\n </m:mathPr></w:WordDocument>\n</xml><![endif]--><!--[if gte mso 9]><xml>\n <w:LatentStyles DefLockedState=\"false\" DefUnhideWhenUsed=\"true\"\n DefSemiHidden=\"true\" DefQFormat=\"false\" DefPriority=\"99\"\n LatentStyleCount=\"276\">\n <w:LsdException Locked=\"false\" Priority=\"0\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"Normal\"/>\n <w:LsdException Locked=\"false\" Priority=\"9\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"heading 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"9\" QFormat=\"true\" Name=\"heading 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"9\" QFormat=\"true\" Name=\"heading 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"9\" QFormat=\"true\" Name=\"heading 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"9\" QFormat=\"true\" Name=\"heading 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"9\" QFormat=\"true\" Name=\"heading 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"9\" QFormat=\"true\" Name=\"heading 7\"/>\n <w:LsdException Locked=\"false\" Priority=\"9\" QFormat=\"true\" Name=\"heading 8\"/>\n <w:LsdException Locked=\"false\" Priority=\"9\" QFormat=\"true\" Name=\"heading 9\"/>\n <w:LsdException Locked=\"false\" Priority=\"39\" Name=\"toc 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"39\" Name=\"toc 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"39\" Name=\"toc 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"39\" Name=\"toc 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"39\" Name=\"toc 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"39\" Name=\"toc 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"39\" Name=\"toc 7\"/>\n <w:LsdException Locked=\"false\" Priority=\"39\" Name=\"toc 8\"/>\n <w:LsdException Locked=\"false\" Priority=\"39\" Name=\"toc 9\"/>\n <w:LsdException Locked=\"false\" Priority=\"35\" QFormat=\"true\" Name=\"caption\"/>\n <w:LsdException Locked=\"false\" Priority=\"10\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"Title\"/>\n <w:LsdException Locked=\"false\" Priority=\"1\" Name=\"Default Paragraph Font\"/>\n <w:LsdException Locked=\"false\" Priority=\"11\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"Subtitle\"/>\n <w:LsdException Locked=\"false\" Priority=\"22\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"Strong\"/>\n <w:LsdException Locked=\"false\" Priority=\"20\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"Emphasis\"/>\n <w:LsdException Locked=\"false\" Priority=\"59\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Table Grid\"/>\n <w:LsdException Locked=\"false\" UnhideWhenUsed=\"false\" Name=\"Placeholder Text\"/>\n <w:LsdException Locked=\"false\" Priority=\"1\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"No Spacing\"/>\n <w:LsdException Locked=\"false\" Priority=\"60\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Shading\"/>\n <w:LsdException Locked=\"false\" Priority=\"61\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light List\"/>\n <w:LsdException Locked=\"false\" Priority=\"62\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Grid\"/>\n <w:LsdException Locked=\"false\" Priority=\"63\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"64\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"65\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"66\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"67\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"68\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"69\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"70\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Dark List\"/>\n <w:LsdException Locked=\"false\" Priority=\"71\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Shading\"/>\n <w:LsdException Locked=\"false\" Priority=\"72\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful List\"/>\n <w:LsdException Locked=\"false\" Priority=\"73\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Grid\"/>\n <w:LsdException Locked=\"false\" Priority=\"60\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Shading Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"61\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light List Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"62\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Grid Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"63\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 1 Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"64\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 2 Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"65\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 1 Accent 1\"/>\n <w:LsdException Locked=\"false\" UnhideWhenUsed=\"false\" Name=\"Revision\"/>\n <w:LsdException Locked=\"false\" Priority=\"34\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"List Paragraph\"/>\n <w:LsdException Locked=\"false\" Priority=\"29\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"Quote\"/>\n <w:LsdException Locked=\"false\" Priority=\"30\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"Intense Quote\"/>\n <w:LsdException Locked=\"false\" Priority=\"66\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 2 Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"67\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 1 Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"68\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 2 Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"69\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 3 Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"70\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Dark List Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"71\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Shading Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"72\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful List Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"73\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Grid Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"60\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Shading Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"61\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light List Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"62\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Grid Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"63\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 1 Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"64\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 2 Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"65\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 1 Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"66\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 2 Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"67\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 1 Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"68\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 2 Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"69\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 3 Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"70\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Dark List Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"71\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Shading Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"72\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful List Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"73\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Grid Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"60\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Shading Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"61\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light List Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"62\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Grid Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"63\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 1 Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"64\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 2 Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"65\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 1 Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"66\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 2 Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"67\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 1 Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"68\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 2 Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"69\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 3 Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"70\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Dark List Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"71\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Shading Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"72\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful List Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"73\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Grid Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"60\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Shading Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"61\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light List Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"62\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Grid Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"63\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 1 Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"64\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 2 Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"65\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 1 Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"66\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 2 Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"67\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 1 Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"68\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 2 Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"69\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 3 Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"70\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Dark List Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"71\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Shading Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"72\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful List Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"73\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Grid Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"60\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Shading Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"61\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light List Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"62\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Grid Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"63\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 1 Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"64\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 2 Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"65\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 1 Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"66\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 2 Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"67\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 1 Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"68\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 2 Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"69\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 3 Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"70\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Dark List Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"71\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Shading Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"72\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful List Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"73\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Grid Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"60\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Shading Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"61\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light List Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"62\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Grid Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"63\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 1 Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"64\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 2 Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"65\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 1 Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"66\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 2 Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"67\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 1 Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"68\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 2 Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"69\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 3 Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"70\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Dark List Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"71\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Shading Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"72\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful List Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"73\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Grid Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"19\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"Subtle Emphasis\"/>\n <w:LsdException Locked=\"false\" Priority=\"21\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"Intense Emphasis\"/>\n <w:LsdException Locked=\"false\" Priority=\"31\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"Subtle Reference\"/>\n <w:LsdException Locked=\"false\" Priority=\"32\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"Intense Reference\"/>\n <w:LsdException Locked=\"false\" Priority=\"33\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"Book Title\"/>\n <w:LsdException Locked=\"false\" Priority=\"37\" Name=\"Bibliography\"/>\n <w:LsdException Locked=\"false\" Priority=\"39\" QFormat=\"true\" Name=\"TOC Heading\"/>\n </w:LatentStyles>\n</xml><![endif]-->\n\n<!--[if gte mso 10]>\n<style>\n /* Style Definitions */\ntable.MsoNormalTable\n {mso-style-name:\"Table Normal\";\n mso-tstyle-rowband-size:0;\n mso-tstyle-colband-size:0;\n mso-style-noshow:yes;\n mso-style-priority:99;\n mso-style-parent:\"\";\n mso-padding-alt:0in 5.4pt 0in 5.4pt;\n mso-para-margin:0in;\n mso-para-margin-bottom:.0001pt;\n mso-pagination:widow-orphan;\n font-size:11.0pt;\n font-family:Calibri;\n mso-fareast-language:ZH-CN;}\n</style>\n<![endif]-->\n\n\n\n<!--StartFragment-->\n\n<div style=\"mso-element:para-border-div;border:solid windowtext 1.0pt;\nmso-border-alt:solid windowtext .5pt;padding:1.0pt 4.0pt 1.0pt 4.0pt\">\n\n<p class=\"MsoNormal\" style=\"margin-bottom:0in;margin-bottom:.0001pt;tab-stops:\n28.35pt;border:none;mso-border-alt:solid windowtext .5pt;padding:0in;\nmso-padding-alt:1.0pt 4.0pt 1.0pt 4.0pt\"><!--[if supportFields]><span\nstyle='font-size:12.0pt;line-height:115%;font-family:Arial'><span\nstyle='mso-element:field-begin'></span><span\nstyle=\"mso-spacerun:yes\"> </span>FORMTEXT <span style='mso-element:field-separator'></span></span><![endif]--><span style=\"font-size:12.0pt;line-height:115%;font-family:Arial\">A</span><span style=\"font-size:12.0pt;line-height:115%;font-family:&quot;Times New Roman&quot;;\nmso-no-proof:yes\">cute focal symptoms from the brain of presumed vascular\norigin have lasted 24 hours or more</span><span style=\"font-size:12.0pt;\nline-height:115%;font-family:Arial\"><!--[if gte mso 9]><xml>\n <w:data>FFFFFFFF00000000140006005400650078007400320031000000000000000000000000000000000000000000000000000000</w:data>\n</xml><![endif]--></span><!--[if supportFields]><span style='font-size:12.0pt;\nline-height:115%;font-family:Arial'><span style='mso-element:field-end'></span></span><![endif]--><span style=\"font-size:12.0pt;line-height:115%;font-family:Arial\"><b><o:p></o:p></b></span></p>\n\n</div>\n\n<!--EndFragment-->"
#> 
#> 
#> $title
#> $title$`@language`
#> [1] "en"
#> 
#> $title$`@value`
#> [1] "Stroke not known if ischaemic or haemorrhagic"
#> 
#> 
#> $definition
#> $definition$`@language`
#> [1] "en"
#> 
#> $definition$`@value`
#> [1] "Fulfills criteria for stroke in acute symptoms of focal brain injury that have lasted 24 hours or more (or led to death before 24 hours), but subtype of stroke (ischemic or haemorrhagic) has not been determined by neuroimaging or other techniques."
#> 
#> 
#> $exclusion
#> $exclusion[[1]]
#> $exclusion[[1]]$label
#> $exclusion[[1]]$label$`@language`
#> [1] "en"
#> 
#> $exclusion[[1]]$label$`@value`
#> [1] "sequelae of stroke"
#> 
#> 
#> $exclusion[[1]]$foundationReference
#> [1] "http://id.who.int/icd/entity/557175275"
#> 
#> $exclusion[[1]]$linearizationReference
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/557175275"
#> 
#> 
#> 
#> $relatedEntitiesInMaternalChapter
#> $relatedEntitiesInMaternalChapter[[1]]
#> [1] "http://id.who.int/icd/entity/564191997"
#> 
#> 
#> $indexTerm
#> $indexTerm[[1]]
#> $indexTerm[[1]]$label
#> $indexTerm[[1]]$label$`@language`
#> [1] "en"
#> 
#> $indexTerm[[1]]$label$`@value`
#> [1] "Stroke not known if ischaemic or haemorrhagic"
#> 
#> 
#> 
#> $indexTerm[[2]]
#> $indexTerm[[2]]$label
#> $indexTerm[[2]]$label$`@language`
#> [1] "en"
#> 
#> $indexTerm[[2]]$label$`@value`
#> [1] "apoplexy"
#> 
#> 
#> 
#> $indexTerm[[3]]
#> $indexTerm[[3]]$label
#> $indexTerm[[3]]$label$`@language`
#> [1] "en"
#> 
#> $indexTerm[[3]]$label$`@value`
#> [1] "brain vascular accident"
#> 
#> 
#> 
#> $indexTerm[[4]]
#> $indexTerm[[4]]$label
#> $indexTerm[[4]]$label$`@language`
#> [1] "en"
#> 
#> $indexTerm[[4]]$label$`@value`
#> [1] "cerebral accident"
#> 
#> 
#> 
#> $indexTerm[[5]]
#> $indexTerm[[5]]$label
#> $indexTerm[[5]]$label$`@language`
#> [1] "en"
#> 
#> $indexTerm[[5]]$label$`@value`
#> [1] "cerebral apoplexy"
#> 
#> 
#> 
#> $indexTerm[[6]]
#> $indexTerm[[6]]$label
#> $indexTerm[[6]]$label$`@language`
#> [1] "en"
#> 
#> $indexTerm[[6]]$label$`@value`
#> [1] "cerebrovascular accident"
#> 
#> 
#> 
#> $indexTerm[[7]]
#> $indexTerm[[7]]$label
#> $indexTerm[[7]]$label$`@language`
#> [1] "en"
#> 
#> $indexTerm[[7]]$label$`@value`
#> [1] "stroke"
#> 
#> 
#> 
#> $indexTerm[[8]]
#> $indexTerm[[8]]$label
#> $indexTerm[[8]]$label$`@language`
#> [1] "en"
#> 
#> $indexTerm[[8]]$label$`@value`
#> [1] "stroke syndrome"
#> 
#> 
#> 
#> $indexTerm[[9]]
#> $indexTerm[[9]]$label
#> $indexTerm[[9]]$label$`@language`
#> [1] "en"
#> 
#> $indexTerm[[9]]$label$`@value`
#> [1] "stroke, not specified as haemorrhage or infarction"
#> 
#> 
#> 
#> $indexTerm[[10]]
#> $indexTerm[[10]]$label
#> $indexTerm[[10]]$label$`@language`
#> [1] "en"
#> 
#> $indexTerm[[10]]$label$`@value`
#> [1] "CVA - [cerebrovascular accident]"
#> 
#> 
#> 
#> $indexTerm[[11]]
#> $indexTerm[[11]]$label
#> $indexTerm[[11]]$label$`@language`
#> [1] "en"
#> 
#> $indexTerm[[11]]$label$`@value`
#> [1] "cerebral vascular accident"
#> 
#> 
#> 
#> $indexTerm[[12]]
#> $indexTerm[[12]]$label
#> $indexTerm[[12]]$label$`@language`
#> [1] "en"
#> 
#> $indexTerm[[12]]$label$`@value`
#> [1] "stroke NOS"
#> 
#> 
#> 
#> $indexTerm[[13]]
#> $indexTerm[[13]]$label
#> $indexTerm[[13]]$label$`@language`
#> [1] "en"
#> 
#> $indexTerm[[13]]$label$`@value`
#> [1] "cerebrovascular incident"
#> 
#> 
#> 
#> $indexTerm[[14]]
#> $indexTerm[[14]]$label
#> $indexTerm[[14]]$label$`@language`
#> [1] "en"
#> 
#> $indexTerm[[14]]$label$`@value`
#> [1] "acute cerebrovascular accident"
#> 
#> 
#> 
#> $indexTerm[[15]]
#> $indexTerm[[15]]$label
#> $indexTerm[[15]]$label$`@language`
#> [1] "en"
#> 
#> $indexTerm[[15]]$label$`@value`
#> [1] "cerebrovascular event"
#> 
#> 
#> 
#> $indexTerm[[16]]
#> $indexTerm[[16]]$label
#> $indexTerm[[16]]$label$`@language`
#> [1] "en"
#> 
#> $indexTerm[[16]]$label$`@value`
#> [1] "massive cerebral vascular accident"
#> 
#> 
#> 
#> $indexTerm[[17]]
#> $indexTerm[[17]]$label
#> $indexTerm[[17]]$label$`@language`
#> [1] "en"
#> 
#> $indexTerm[[17]]$label$`@value`
#> [1] "massive cerebrovascular accident"
#> 
#> 
#> 
#> 
# Assuming that the CLIENT ID and CLIENT SECRET are set up. Substitute accordingly
if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {
  # Generated token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)

  # Search for ICD-11 URI across all linearizations in a specific release
  icd11_id2(token, id = "http://id.who.int/icd/entity/1442995018")

  # Search for ICD-11 URI with parents and diagnostic criteria across all linearizations
  #  in a specific release
  icd11_id2(token,
    id = "http://id.who.int/icd/entity/1442995018",
    include = c("ancestor", "diagnosticCriteria")
  )
}
#> $`@context`
#> [1] "http://id.who.int/icd/contexts/contextForFoundationEntity.json"
#> 
#> $`@id`
#> [1] "http://id.who.int/icd/entity/1442995018"
#> 
#> $parent
#> $parent[[1]]
#> [1] "http://id.who.int/icd/entity/843843448"
#> 
#> 
#> $browserUrl
#> [1] "https://icd.who.int/browse/2024-01/foundation/en#1442995018"
#> 
#> $ancestor
#> $ancestor[[1]]
#> [1] "http://id.who.int/icd/entity/843843448"
#> 
#> $ancestor[[2]]
#> [1] "http://id.who.int/icd/entity/1296093776"
#> 
#> $ancestor[[3]]
#> [1] "http://id.who.int/icd/entity/455013390"
#> 
#> $ancestor[[4]]
#> [1] "http://id.who.int/icd/entity/448895267"
#> 
#> $ancestor[[5]]
#> [1] "http://id.who.int/icd/entity/426429380"
#> 
#> 
#> $diagnosticCriteria
#> $diagnosticCriteria$`@language`
#> [1] "en"
#> 
#> $diagnosticCriteria$`@value`
#> [1] "Signs and symtoms:&nbsp;\n\n\n\n\n\n\n<!--[if gte mso 9]><xml>\n <o:DocumentProperties>\n <o:Revision>0</o:Revision>\n <o:TotalTime>0</o:TotalTime>\n <o:Pages>1</o:Pages>\n <o:Words>16</o:Words>\n <o:Characters>92</o:Characters>\n <o:Company>University of California, Berkeley</o:Company>\n <o:Lines>1</o:Lines>\n <o:Paragraphs>1</o:Paragraphs>\n <o:CharactersWithSpaces>107</o:CharactersWithSpaces>\n <o:Version>14.0</o:Version>\n </o:DocumentProperties>\n</xml><![endif]-->\n\n<!--[if gte mso 9]><xml>\n <w:WordDocument>\n <w:View>Normal</w:View>\n <w:Zoom>0</w:Zoom>\n <w:TrackMoves/>\n <w:TrackFormatting/>\n <w:PunctuationKerning/>\n <w:ValidateAgainstSchemas/>\n <w:SaveIfXMLInvalid>false</w:SaveIfXMLInvalid>\n <w:IgnoreMixedContent>false</w:IgnoreMixedContent>\n <w:AlwaysShowPlaceholderText>false</w:AlwaysShowPlaceholderText>\n <w:DoNotPromoteQF/>\n <w:LidThemeOther>EN-US</w:LidThemeOther>\n <w:LidThemeAsian>JA</w:LidThemeAsian>\n <w:LidThemeComplexScript>X-NONE</w:LidThemeComplexScript>\n <w:Compatibility>\n <w:BreakWrappedTables/>\n <w:SnapToGridInCell/>\n <w:WrapTextWithPunct/>\n <w:UseAsianBreakRules/>\n <w:DontGrowAutofit/>\n <w:SplitPgBreakAndParaMark/>\n <w:EnableOpenTypeKerning/>\n <w:DontFlipMirrorIndents/>\n <w:OverrideTableStyleHps/>\n </w:Compatibility>\n <m:mathPr>\n <m:mathFont m:val=\"Cambria Math\"/>\n <m:brkBin m:val=\"before\"/>\n <m:brkBinSub m:val=\"&#45;-\"/>\n <m:smallFrac m:val=\"off\"/>\n <m:dispDef/>\n <m:lMargin m:val=\"0\"/>\n <m:rMargin m:val=\"0\"/>\n <m:defJc m:val=\"centerGroup\"/>\n <m:wrapIndent m:val=\"1440\"/>\n <m:intLim m:val=\"subSup\"/>\n <m:naryLim m:val=\"undOvr\"/>\n </m:mathPr></w:WordDocument>\n</xml><![endif]--><!--[if gte mso 9]><xml>\n <w:LatentStyles DefLockedState=\"false\" DefUnhideWhenUsed=\"true\"\n DefSemiHidden=\"true\" DefQFormat=\"false\" DefPriority=\"99\"\n LatentStyleCount=\"276\">\n <w:LsdException Locked=\"false\" Priority=\"0\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"Normal\"/>\n <w:LsdException Locked=\"false\" Priority=\"9\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"heading 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"9\" QFormat=\"true\" Name=\"heading 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"9\" QFormat=\"true\" Name=\"heading 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"9\" QFormat=\"true\" Name=\"heading 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"9\" QFormat=\"true\" Name=\"heading 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"9\" QFormat=\"true\" Name=\"heading 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"9\" QFormat=\"true\" Name=\"heading 7\"/>\n <w:LsdException Locked=\"false\" Priority=\"9\" QFormat=\"true\" Name=\"heading 8\"/>\n <w:LsdException Locked=\"false\" Priority=\"9\" QFormat=\"true\" Name=\"heading 9\"/>\n <w:LsdException Locked=\"false\" Priority=\"39\" Name=\"toc 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"39\" Name=\"toc 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"39\" Name=\"toc 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"39\" Name=\"toc 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"39\" Name=\"toc 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"39\" Name=\"toc 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"39\" Name=\"toc 7\"/>\n <w:LsdException Locked=\"false\" Priority=\"39\" Name=\"toc 8\"/>\n <w:LsdException Locked=\"false\" Priority=\"39\" Name=\"toc 9\"/>\n <w:LsdException Locked=\"false\" Priority=\"35\" QFormat=\"true\" Name=\"caption\"/>\n <w:LsdException Locked=\"false\" Priority=\"10\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"Title\"/>\n <w:LsdException Locked=\"false\" Priority=\"1\" Name=\"Default Paragraph Font\"/>\n <w:LsdException Locked=\"false\" Priority=\"11\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"Subtitle\"/>\n <w:LsdException Locked=\"false\" Priority=\"22\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"Strong\"/>\n <w:LsdException Locked=\"false\" Priority=\"20\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"Emphasis\"/>\n <w:LsdException Locked=\"false\" Priority=\"59\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Table Grid\"/>\n <w:LsdException Locked=\"false\" UnhideWhenUsed=\"false\" Name=\"Placeholder Text\"/>\n <w:LsdException Locked=\"false\" Priority=\"1\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"No Spacing\"/>\n <w:LsdException Locked=\"false\" Priority=\"60\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Shading\"/>\n <w:LsdException Locked=\"false\" Priority=\"61\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light List\"/>\n <w:LsdException Locked=\"false\" Priority=\"62\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Grid\"/>\n <w:LsdException Locked=\"false\" Priority=\"63\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"64\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"65\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"66\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"67\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"68\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"69\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"70\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Dark List\"/>\n <w:LsdException Locked=\"false\" Priority=\"71\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Shading\"/>\n <w:LsdException Locked=\"false\" Priority=\"72\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful List\"/>\n <w:LsdException Locked=\"false\" Priority=\"73\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Grid\"/>\n <w:LsdException Locked=\"false\" Priority=\"60\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Shading Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"61\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light List Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"62\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Grid Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"63\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 1 Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"64\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 2 Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"65\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 1 Accent 1\"/>\n <w:LsdException Locked=\"false\" UnhideWhenUsed=\"false\" Name=\"Revision\"/>\n <w:LsdException Locked=\"false\" Priority=\"34\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"List Paragraph\"/>\n <w:LsdException Locked=\"false\" Priority=\"29\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"Quote\"/>\n <w:LsdException Locked=\"false\" Priority=\"30\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"Intense Quote\"/>\n <w:LsdException Locked=\"false\" Priority=\"66\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 2 Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"67\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 1 Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"68\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 2 Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"69\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 3 Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"70\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Dark List Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"71\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Shading Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"72\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful List Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"73\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Grid Accent 1\"/>\n <w:LsdException Locked=\"false\" Priority=\"60\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Shading Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"61\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light List Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"62\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Grid Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"63\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 1 Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"64\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 2 Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"65\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 1 Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"66\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 2 Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"67\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 1 Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"68\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 2 Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"69\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 3 Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"70\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Dark List Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"71\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Shading Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"72\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful List Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"73\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Grid Accent 2\"/>\n <w:LsdException Locked=\"false\" Priority=\"60\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Shading Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"61\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light List Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"62\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Grid Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"63\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 1 Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"64\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 2 Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"65\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 1 Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"66\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 2 Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"67\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 1 Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"68\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 2 Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"69\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 3 Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"70\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Dark List Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"71\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Shading Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"72\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful List Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"73\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Grid Accent 3\"/>\n <w:LsdException Locked=\"false\" Priority=\"60\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Shading Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"61\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light List Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"62\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Grid Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"63\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 1 Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"64\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 2 Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"65\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 1 Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"66\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 2 Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"67\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 1 Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"68\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 2 Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"69\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 3 Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"70\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Dark List Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"71\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Shading Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"72\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful List Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"73\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Grid Accent 4\"/>\n <w:LsdException Locked=\"false\" Priority=\"60\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Shading Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"61\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light List Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"62\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Grid Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"63\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 1 Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"64\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 2 Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"65\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 1 Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"66\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 2 Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"67\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 1 Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"68\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 2 Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"69\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 3 Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"70\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Dark List Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"71\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Shading Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"72\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful List Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"73\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Grid Accent 5\"/>\n <w:LsdException Locked=\"false\" Priority=\"60\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Shading Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"61\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light List Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"62\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Light Grid Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"63\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 1 Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"64\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Shading 2 Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"65\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 1 Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"66\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium List 2 Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"67\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 1 Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"68\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 2 Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"69\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Medium Grid 3 Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"70\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Dark List Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"71\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Shading Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"72\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful List Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"73\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" Name=\"Colorful Grid Accent 6\"/>\n <w:LsdException Locked=\"false\" Priority=\"19\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"Subtle Emphasis\"/>\n <w:LsdException Locked=\"false\" Priority=\"21\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"Intense Emphasis\"/>\n <w:LsdException Locked=\"false\" Priority=\"31\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"Subtle Reference\"/>\n <w:LsdException Locked=\"false\" Priority=\"32\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"Intense Reference\"/>\n <w:LsdException Locked=\"false\" Priority=\"33\" SemiHidden=\"false\"\n UnhideWhenUsed=\"false\" QFormat=\"true\" Name=\"Book Title\"/>\n <w:LsdException Locked=\"false\" Priority=\"37\" Name=\"Bibliography\"/>\n <w:LsdException Locked=\"false\" Priority=\"39\" QFormat=\"true\" Name=\"TOC Heading\"/>\n </w:LatentStyles>\n</xml><![endif]-->\n\n<!--[if gte mso 10]>\n<style>\n /* Style Definitions */\ntable.MsoNormalTable\n {mso-style-name:\"Table Normal\";\n mso-tstyle-rowband-size:0;\n mso-tstyle-colband-size:0;\n mso-style-noshow:yes;\n mso-style-priority:99;\n mso-style-parent:\"\";\n mso-padding-alt:0in 5.4pt 0in 5.4pt;\n mso-para-margin:0in;\n mso-para-margin-bottom:.0001pt;\n mso-pagination:widow-orphan;\n font-size:11.0pt;\n font-family:Calibri;\n mso-fareast-language:ZH-CN;}\n</style>\n<![endif]-->\n\n\n\n<!--StartFragment-->\n\n<div style=\"mso-element:para-border-div;border:solid windowtext 1.0pt;\nmso-border-alt:solid windowtext .5pt;padding:1.0pt 4.0pt 1.0pt 4.0pt\">\n\n<p class=\"MsoNormal\" style=\"margin-bottom:0in;margin-bottom:.0001pt;tab-stops:\n28.35pt;border:none;mso-border-alt:solid windowtext .5pt;padding:0in;\nmso-padding-alt:1.0pt 4.0pt 1.0pt 4.0pt\"><!--[if supportFields]><span\nstyle='font-size:12.0pt;line-height:115%;font-family:Arial'><span\nstyle='mso-element:field-begin'></span><span\nstyle=\"mso-spacerun:yes\"> </span>FORMTEXT <span style='mso-element:field-separator'></span></span><![endif]--><span style=\"font-size:12.0pt;line-height:115%;font-family:Arial\">A</span><span style=\"font-size:12.0pt;line-height:115%;font-family:&quot;Times New Roman&quot;;\nmso-no-proof:yes\">cute focal symptoms from the brain of presumed vascular\norigin have lasted 24 hours or more</span><span style=\"font-size:12.0pt;\nline-height:115%;font-family:Arial\"><!--[if gte mso 9]><xml>\n <w:data>FFFFFFFF00000000140006005400650078007400320031000000000000000000000000000000000000000000000000000000</w:data>\n</xml><![endif]--></span><!--[if supportFields]><span style='font-size:12.0pt;\nline-height:115%;font-family:Arial'><span style='mso-element:field-end'></span></span><![endif]--><span style=\"font-size:12.0pt;line-height:115%;font-family:Arial\"><b><o:p></o:p></b></span></p>\n\n</div>\n\n<!--EndFragment-->"
#> 
#> 
#> $title
#> $title$`@language`
#> [1] "en"
#> 
#> $title$`@value`
#> [1] "Stroke not known if ischaemic or haemorrhagic"
#> 
#> 
#> $synonym
#> $synonym[[1]]
#> $synonym[[1]]$label
#> $synonym[[1]]$label$`@language`
#> [1] "en"
#> 
#> $synonym[[1]]$label$`@value`
#> [1] "apoplexy"
#> 
#> 
#> 
#> $synonym[[2]]
#> $synonym[[2]]$label
#> $synonym[[2]]$label$`@language`
#> [1] "en"
#> 
#> $synonym[[2]]$label$`@value`
#> [1] "brain vascular accident"
#> 
#> 
#> 
#> $synonym[[3]]
#> $synonym[[3]]$label
#> $synonym[[3]]$label$`@language`
#> [1] "en"
#> 
#> $synonym[[3]]$label$`@value`
#> [1] "cerebral accident"
#> 
#> 
#> 
#> $synonym[[4]]
#> $synonym[[4]]$label
#> $synonym[[4]]$label$`@language`
#> [1] "en"
#> 
#> $synonym[[4]]$label$`@value`
#> [1] "cerebral apoplexy"
#> 
#> 
#> 
#> $synonym[[5]]
#> $synonym[[5]]$label
#> $synonym[[5]]$label$`@language`
#> [1] "en"
#> 
#> $synonym[[5]]$label$`@value`
#> [1] "cerebrovascular accident"
#> 
#> 
#> 
#> $synonym[[6]]
#> $synonym[[6]]$label
#> $synonym[[6]]$label$`@language`
#> [1] "en"
#> 
#> $synonym[[6]]$label$`@value`
#> [1] "stroke"
#> 
#> 
#> 
#> $synonym[[7]]
#> $synonym[[7]]$label
#> $synonym[[7]]$label$`@language`
#> [1] "en"
#> 
#> $synonym[[7]]$label$`@value`
#> [1] "stroke syndrome"
#> 
#> 
#> 
#> $synonym[[8]]
#> $synonym[[8]]$label
#> $synonym[[8]]$label$`@language`
#> [1] "en"
#> 
#> $synonym[[8]]$label$`@value`
#> [1] "stroke, not specified as haemorrhage or infarction"
#> 
#> 
#> 
#> $synonym[[9]]
#> $synonym[[9]]$label
#> $synonym[[9]]$label$`@language`
#> [1] "en"
#> 
#> $synonym[[9]]$label$`@value`
#> [1] "CVA - [cerebrovascular accident]"
#> 
#> 
#> 
#> $synonym[[10]]
#> $synonym[[10]]$label
#> $synonym[[10]]$label$`@language`
#> [1] "en"
#> 
#> $synonym[[10]]$label$`@value`
#> [1] "cerebral vascular accident"
#> 
#> 
#> 
#> $synonym[[11]]
#> $synonym[[11]]$label
#> $synonym[[11]]$label$`@language`
#> [1] "en"
#> 
#> $synonym[[11]]$label$`@value`
#> [1] "stroke NOS"
#> 
#> 
#> 
#> $synonym[[12]]
#> $synonym[[12]]$label
#> $synonym[[12]]$label$`@language`
#> [1] "en"
#> 
#> $synonym[[12]]$label$`@value`
#> [1] "cerebrovascular incident"
#> 
#> 
#> 
#> $synonym[[13]]
#> $synonym[[13]]$label
#> $synonym[[13]]$label$`@language`
#> [1] "en"
#> 
#> $synonym[[13]]$label$`@value`
#> [1] "acute cerebrovascular accident"
#> 
#> 
#> 
#> $synonym[[14]]
#> $synonym[[14]]$label
#> $synonym[[14]]$label$`@language`
#> [1] "en"
#> 
#> $synonym[[14]]$label$`@value`
#> [1] "cerebrovascular event"
#> 
#> 
#> 
#> $synonym[[15]]
#> $synonym[[15]]$label
#> $synonym[[15]]$label$`@language`
#> [1] "en"
#> 
#> $synonym[[15]]$label$`@value`
#> [1] "massive cerebral vascular accident"
#> 
#> 
#> 
#> $synonym[[16]]
#> $synonym[[16]]$label
#> $synonym[[16]]$label$`@language`
#> [1] "en"
#> 
#> $synonym[[16]]$label$`@value`
#> [1] "massive cerebrovascular accident"
#> 
#> 
#> 
#> 
#> $definition
#> $definition$`@language`
#> [1] "en"
#> 
#> $definition$`@value`
#> [1] "Fulfills criteria for stroke in acute symptoms of focal brain injury that have lasted 24 hours or more (or led to death before 24 hours), but subtype of stroke (ischemic or haemorrhagic) has not been determined by neuroimaging or other techniques."
#> 
#> 
#> $exclusion
#> $exclusion[[1]]
#> $exclusion[[1]]$label
#> $exclusion[[1]]$label$`@language`
#> [1] "en"
#> 
#> $exclusion[[1]]$label$`@value`
#> [1] "sequelae of stroke"
#> 
#> 
#> $exclusion[[1]]$foundationReference
#> [1] "http://id.who.int/icd/entity/557175275"
#> 
#> 
#> 
#> $relatedEntitiesInMaternalChapter
#> $relatedEntitiesInMaternalChapter[[1]]
#> [1] "http://id.who.int/icd/entity/564191997"
#> 
#> 
# Assuming that the CLIENT ID and CLIENT SECRET are set up. Substitute accordingly
if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {
  # Generated token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)

  # Search for ICD-11 URI across multiple releases
  icd11_id3(token, id = "http://id.who.int/icd/entity/1442995018")

}
#> $`@context`
#> [1] "http://id.who.int/icd/contexts/contextForMultiVersion.json"
#> 
#> $`@id`
#> [1] "http://id.who.int/icd/release/11/mms/1442995018"
#> 
#> $title
#> $title$`@language`
#> [1] "en"
#> 
#> $title$`@value`
#> [1] "Stroke not known if ischaemic or haemorrhagic"
#> 
#> 
#> $latestRelease
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1442995018"
#> 
#> $release
#> $release[[1]]
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1442995018"
#> 
#> $release[[2]]
#> [1] "http://id.who.int/icd/release/11/2023-01/mms/1442995018"
#> 
#> $release[[3]]
#> [1] "http://id.who.int/icd/release/11/2022-02/mms/1442995018"
#> 
#> $release[[4]]
#> [1] "http://id.who.int/icd/release/11/2021-05/mms/1442995018"
#> 
#> $release[[5]]
#> [1] "http://id.who.int/icd/release/11/2020-09/mms/1442995018"
#> 
#> $release[[6]]
#> [1] "http://id.who.int/icd/release/11/2019-04/mms/1442995018"
#> 
#> $release[[7]]
#> [1] "http://id.who.int/icd/release/11/2018/mms/1442995018"
#> 
#>