Search

Searches the foundation component of the ICD-11.

icd11_search(
  token,
  text,
  release = "2024-01",
  linearization = "mms",
  subtrees_filter = NULL,
  subtrees_filter_foundation_descendents = FALSE,
  include_keyword_result = FALSE,
  chapter_filter = NULL,
  use_flexisearch = FALSE,
  flat_results = NULL,
  highlighting_enabled = FALSE,
  medical_coding_mode = TRUE,
  properties_to_be_searched = NULL,
  language = "en",
  auto_update = TRUE
)

icd11_search2(
  token,
  text,
  release = "2024-01",
  linearization = "mms",
  subtrees_filter = NULL,
  chapter_filter = NULL,
  use_flexisearch = FALSE,
  flat_results = NULL,
  highlighting_enabled = FALSE,
  properties_to_be_searched = NULL,
  language = "en",
  auto_update = TRUE
)

Arguments

token

(list) Access token to the API obtained from get_token()

text

The text component to search in ICD-11.

release

Release version as listed in https://icd.who.int/docs/icd-api/SupportedClassifications/#supported-classifications-and-versions the default is 2024-01.

linearization

(character) ICD-11 linearization (default = "mms"). See all linearizations in the WHO's API website https://icd.who.int/docs/icd-api/APIDoc-Version2/#the-icd11-linearizations

subtrees_filter

(string) Optional parameter. Vector of URIs. If provided, the search will be performed on the entities provided and their descendants

subtrees_filter_foundation_descendents

When a subtrees_filter is used the search looks at the linearization descendants of the entities in the subtrees_filter. If

subtrees_filter_foundation_descendents is set to TRUE, the search looks at the foundation descendants of the entities in the subtrees_filter. Default value: FALSE

include_keyword_result

Boolean. Optional parameter. If set to TRUE, the search result will also include keyword list. Default value: FALSE

chapter_filter

Vector of chapter codes e.g., c("01","02","21"). When provided, the search will be performed only on these chapters.

use_flexisearch

Boolean. Changes the search mode to flexible search. Default value: FALSE

flat_results

Boolean. Optional parameter. If set to true the search result entities are provided in a nested data structure representing the ICD-11 hierarchy. Default value: true.

highlighting_enabled

Boolean. Optional. If set to FALSE the search result highlighting is turned off. Default value: FALSE. Highlighting is in html format.

medical_coding_mode

Boolean. This mode is the default and should be used when searching the classification for medical coding purposes. Default value: TRUE.

properties_to_be_searched

String. The properties to be included in the search. Used only when medical_coding_mode=FALSE

language

(character) Language for the request. Available languages are listed in https://icd.who.int/docs/icd-api/SupportedClassifications/ the default is English (language = "en").

auto_update

Attempts to update token automatically

Value

A list of the coded information as well as any additional information.

Details

The icd11_search function allows you to search across different linearizations such as the Mortality and Morbidity Statistics linearization = "mms" (default) and the International Classifications of Functioning Disability and Health linearization = "icf".

Note

Both icd11_search() and icd11_search2() provide options for search. They correspond to two different API entry-points in the WHO's API and provide slighly different amount of information for the same result.

This provides the default search but might throw different amount of information as icd11_search2.

Examples

# Assuming that the CLIENT ID and CLIENT SECRET are set up. Substitute accordingly
if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {
  # Generated token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)

  # Search for ICD-10 releases
  icd11_search(token, text = "HIV stage 4")
}
#> $error
#> [1] FALSE
#> 
#> $errorMessage
#> NULL
#> 
#> $resultChopped
#> [1] FALSE
#> 
#> $wordSuggestionsChopped
#> [1] FALSE
#> 
#> $guessType
#> [1] 2
#> 
#> $uniqueSearchId
#> [1] "945db7f4-b7c1-46c6-bfd5-e85959eeef06"
#> 
#> $words
#> NULL
#> 
#> $destinationEntities
#> $destinationEntities[[1]]
#> $destinationEntities[[1]]$id
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1280139157/unspecified"
#> 
#> $destinationEntities[[1]]$title
#> [1] "HIV disease clinical stage 4 associated with tuberculosis, unspecified"
#> 
#> $destinationEntities[[1]]$stemId
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1280139157/unspecified"
#> 
#> $destinationEntities[[1]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[1]]$postcoordinationAvailability
#> [1] 1
#> 
#> $destinationEntities[[1]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$hasMaternalChapterLink
#> [1] TRUE
#> 
#> $destinationEntities[[1]]$hasPerinatalChapterLink
#> [1] TRUE
#> 
#> $destinationEntities[[1]]$matchingPVs
#> $destinationEntities[[1]]$matchingPVs[[1]]
#> $destinationEntities[[1]]$matchingPVs[[1]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[1]]$matchingPVs[[1]]$label
#> [1] "HIV disease clinical stage 4 associated with TB"
#> 
#> $destinationEntities[[1]]$matchingPVs[[1]]$score
#> [1] 0.501157
#> 
#> $destinationEntities[[1]]$matchingPVs[[1]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$matchingPVs[[1]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1280139157"
#> 
#> $destinationEntities[[1]]$matchingPVs[[1]]$propertyValueType
#> [1] 0
#> 
#> 
#> $destinationEntities[[1]]$matchingPVs[[2]]
#> $destinationEntities[[1]]$matchingPVs[[2]]$propertyId
#> [1] "TitleOfOtherEntity"
#> 
#> $destinationEntities[[1]]$matchingPVs[[2]]$label
#> [1] "HIV disease clinical stage 4 associated with tuberculosis"
#> 
#> $destinationEntities[[1]]$matchingPVs[[2]]$score
#> [1] 0.326967
#> 
#> $destinationEntities[[1]]$matchingPVs[[2]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$matchingPVs[[2]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1280139157"
#> 
#> $destinationEntities[[1]]$matchingPVs[[2]]$propertyValueType
#> [1] 0
#> 
#> 
#> 
#> $destinationEntities[[1]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$isResidualUnspecified
#> [1] TRUE
#> 
#> $destinationEntities[[1]]$chapter
#> [1] "01"
#> 
#> $destinationEntities[[1]]$theCode
#> [1] "1C60.3Z"
#> 
#> $destinationEntities[[1]]$score
#> [1] 0.501157
#> 
#> $destinationEntities[[1]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[1]]$titleIsTopScore
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[1]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[2]]
#> $destinationEntities[[2]]$id
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1280139157/other"
#> 
#> $destinationEntities[[2]]$title
#> [1] "Other specified HIV disease clinical stage 4 associated with tuberculosis"
#> 
#> $destinationEntities[[2]]$stemId
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1280139157/other"
#> 
#> $destinationEntities[[2]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[2]]$postcoordinationAvailability
#> [1] 1
#> 
#> $destinationEntities[[2]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[2]]$hasMaternalChapterLink
#> [1] TRUE
#> 
#> $destinationEntities[[2]]$hasPerinatalChapterLink
#> [1] TRUE
#> 
#> $destinationEntities[[2]]$matchingPVs
#> list()
#> 
#> $destinationEntities[[2]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[2]]$isResidualOther
#> [1] TRUE
#> 
#> $destinationEntities[[2]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[2]]$chapter
#> [1] "01"
#> 
#> $destinationEntities[[2]]$theCode
#> [1] "1C60.3Y"
#> 
#> $destinationEntities[[2]]$score
#> [1] 0.2530371
#> 
#> $destinationEntities[[2]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[2]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[2]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[2]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[2]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[3]]
#> $destinationEntities[[3]]$id
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/97626683/unspecified"
#> 
#> $destinationEntities[[3]]$title
#> [1] "HIV disease clinical stage 4 associated with malaria, unspecified"
#> 
#> $destinationEntities[[3]]$stemId
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/97626683/unspecified"
#> 
#> $destinationEntities[[3]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[3]]$postcoordinationAvailability
#> [1] 1
#> 
#> $destinationEntities[[3]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$hasMaternalChapterLink
#> [1] TRUE
#> 
#> $destinationEntities[[3]]$hasPerinatalChapterLink
#> [1] TRUE
#> 
#> $destinationEntities[[3]]$matchingPVs
#> $destinationEntities[[3]]$matchingPVs[[1]]
#> $destinationEntities[[3]]$matchingPVs[[1]]$propertyId
#> [1] "TitleOfOtherEntity"
#> 
#> $destinationEntities[[3]]$matchingPVs[[1]]$label
#> [1] "HIV disease clinical stage 4 associated with malaria"
#> 
#> $destinationEntities[[3]]$matchingPVs[[1]]$score
#> [1] 0.1622578
#> 
#> $destinationEntities[[3]]$matchingPVs[[1]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$matchingPVs[[1]]$foundationUri
#> [1] "http://id.who.int/icd/entity/97626683"
#> 
#> $destinationEntities[[3]]$matchingPVs[[1]]$propertyValueType
#> [1] 0
#> 
#> 
#> 
#> $destinationEntities[[3]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$isResidualUnspecified
#> [1] TRUE
#> 
#> $destinationEntities[[3]]$chapter
#> [1] "01"
#> 
#> $destinationEntities[[3]]$theCode
#> [1] "1C61.3Z"
#> 
#> $destinationEntities[[3]]$score
#> [1] 0.1622578
#> 
#> $destinationEntities[[3]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[3]]$titleIsTopScore
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[3]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[4]]
#> $destinationEntities[[4]]$id
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/97626683/other"
#> 
#> $destinationEntities[[4]]$title
#> [1] "Other specified HIV disease clinical stage 4 associated with malaria"
#> 
#> $destinationEntities[[4]]$stemId
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/97626683/other"
#> 
#> $destinationEntities[[4]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[4]]$postcoordinationAvailability
#> [1] 1
#> 
#> $destinationEntities[[4]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$hasMaternalChapterLink
#> [1] TRUE
#> 
#> $destinationEntities[[4]]$hasPerinatalChapterLink
#> [1] TRUE
#> 
#> $destinationEntities[[4]]$matchingPVs
#> list()
#> 
#> $destinationEntities[[4]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$isResidualOther
#> [1] TRUE
#> 
#> $destinationEntities[[4]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$chapter
#> [1] "01"
#> 
#> $destinationEntities[[4]]$theCode
#> [1] "1C61.3Y"
#> 
#> $destinationEntities[[4]]$score
#> [1] 0.08832788
#> 
#> $destinationEntities[[4]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[4]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[4]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[4]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[5]]
#> $destinationEntities[[5]]$id
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1858812010/unspecified"
#> 
#> $destinationEntities[[5]]$title
#> [1] "HIV disease clinical stage 4 without mention of tuberculosis or malaria, unspecified"
#> 
#> $destinationEntities[[5]]$stemId
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1858812010/unspecified"
#> 
#> $destinationEntities[[5]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[5]]$postcoordinationAvailability
#> [1] 1
#> 
#> $destinationEntities[[5]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[5]]$hasMaternalChapterLink
#> [1] TRUE
#> 
#> $destinationEntities[[5]]$hasPerinatalChapterLink
#> [1] TRUE
#> 
#> $destinationEntities[[5]]$matchingPVs
#> $destinationEntities[[5]]$matchingPVs[[1]]
#> $destinationEntities[[5]]$matchingPVs[[1]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[5]]$matchingPVs[[1]]$label
#> [1] "HIV disease clinical stage 4 without tuberculosis or malaria"
#> 
#> $destinationEntities[[5]]$matchingPVs[[1]]$score
#> [1] -0.249582
#> 
#> $destinationEntities[[5]]$matchingPVs[[1]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[5]]$matchingPVs[[1]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1858812010"
#> 
#> $destinationEntities[[5]]$matchingPVs[[1]]$propertyValueType
#> [1] 0
#> 
#> 
#> $destinationEntities[[5]]$matchingPVs[[2]]
#> $destinationEntities[[5]]$matchingPVs[[2]]$propertyId
#> [1] "TitleOfOtherEntity"
#> 
#> $destinationEntities[[5]]$matchingPVs[[2]]$label
#> [1] "HIV disease clinical stage 4 without mention of tuberculosis or malaria"
#> 
#> $destinationEntities[[5]]$matchingPVs[[2]]$score
#> [1] -0.4095458
#> 
#> $destinationEntities[[5]]$matchingPVs[[2]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[5]]$matchingPVs[[2]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1858812010"
#> 
#> $destinationEntities[[5]]$matchingPVs[[2]]$propertyValueType
#> [1] 0
#> 
#> 
#> 
#> $destinationEntities[[5]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[5]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[5]]$isResidualUnspecified
#> [1] TRUE
#> 
#> $destinationEntities[[5]]$chapter
#> [1] "01"
#> 
#> $destinationEntities[[5]]$theCode
#> [1] "1C62.3Z"
#> 
#> $destinationEntities[[5]]$score
#> [1] -0.249582
#> 
#> $destinationEntities[[5]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[5]]$titleIsTopScore
#> [1] FALSE
#> 
#> $destinationEntities[[5]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[5]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[5]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[6]]
#> $destinationEntities[[6]]$id
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1173624218 / http://id.who.int/icd/release/11/2024-01/mms/1280139157/unspecified"
#> 
#> $destinationEntities[[6]]$title
#> [1] "Dementia due to human immunodeficiency virus [HIV disease clinical stage 4 associated with tuberculosis, unspecified]"
#> 
#> $destinationEntities[[6]]$stemId
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1173624218"
#> 
#> $destinationEntities[[6]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[6]]$postcoordinationAvailability
#> [1] 2
#> 
#> $destinationEntities[[6]]$hasCodingNote
#> [1] TRUE
#> 
#> $destinationEntities[[6]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[6]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[6]]$matchingPVs
#> $destinationEntities[[6]]$matchingPVs[[1]]
#> $destinationEntities[[6]]$matchingPVs[[1]]$propertyId
#> [1] "Title"
#> 
#> $destinationEntities[[6]]$matchingPVs[[1]]$label
#> [1] "Dementia due to human immunodeficiency virus[HIV disease clinical stage 4 associated with TB]"
#> 
#> $destinationEntities[[6]]$matchingPVs[[1]]$score
#> [1] -0.3301043
#> 
#> $destinationEntities[[6]]$matchingPVs[[1]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[6]]$matchingPVs[[1]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/1280139157"
#> 
#> $destinationEntities[[6]]$matchingPVs[[1]]$propertyValueType
#> [1] 3
#> 
#> 
#> $destinationEntities[[6]]$matchingPVs[[2]]
#> $destinationEntities[[6]]$matchingPVs[[2]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[6]]$matchingPVs[[2]]$label
#> [1] "HIV - [human immunodeficiency virus] dementia[HIV disease clinical stage 4 associated with TB]"
#> 
#> $destinationEntities[[6]]$matchingPVs[[2]]$score
#> [1] -0.3301043
#> 
#> $destinationEntities[[6]]$matchingPVs[[2]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[6]]$matchingPVs[[2]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/1280139157"
#> 
#> $destinationEntities[[6]]$matchingPVs[[2]]$propertyValueType
#> [1] 3
#> 
#> 
#> $destinationEntities[[6]]$matchingPVs[[3]]
#> $destinationEntities[[6]]$matchingPVs[[3]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[6]]$matchingPVs[[3]]$label
#> [1] "HIV- [human immunodeficiency virus] associated cognitive motor complex[HIV disease clinical stage 4 associated with TB]"
#> 
#> $destinationEntities[[6]]$matchingPVs[[3]]$score
#> [1] -0.334101
#> 
#> $destinationEntities[[6]]$matchingPVs[[3]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[6]]$matchingPVs[[3]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/1280139157"
#> 
#> $destinationEntities[[6]]$matchingPVs[[3]]$propertyValueType
#> [1] 3
#> 
#> 
#> $destinationEntities[[6]]$matchingPVs[[4]]
#> $destinationEntities[[6]]$matchingPVs[[4]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[6]]$matchingPVs[[4]]$label
#> [1] "HIV - [human immunodeficiency virus] dementia[HIV disease clinical stage 4 associated with tuberculosis, unspecified]"
#> 
#> $destinationEntities[[6]]$matchingPVs[[4]]$score
#> [1] -0.4520373
#> 
#> $destinationEntities[[6]]$matchingPVs[[4]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[6]]$matchingPVs[[4]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/1280139157"
#> 
#> $destinationEntities[[6]]$matchingPVs[[4]]$propertyValueType
#> [1] 3
#> 
#> 
#> $destinationEntities[[6]]$matchingPVs[[5]]
#> $destinationEntities[[6]]$matchingPVs[[5]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[6]]$matchingPVs[[5]]$label
#> [1] "HIV- [human immunodeficiency virus] associated cognitive motor complex[HIV disease clinical stage 4 associated with tuberculosis, unspecified]"
#> 
#> $destinationEntities[[6]]$matchingPVs[[5]]$score
#> [1] -0.4560341
#> 
#> $destinationEntities[[6]]$matchingPVs[[5]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[6]]$matchingPVs[[5]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/1280139157"
#> 
#> $destinationEntities[[6]]$matchingPVs[[5]]$propertyValueType
#> [1] 3
#> 
#> 
#> $destinationEntities[[6]]$matchingPVs[[6]]
#> $destinationEntities[[6]]$matchingPVs[[6]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[6]]$matchingPVs[[6]]$label
#> [1] "HIV- [human immunodeficiency virus] associated dementia complex[HIV disease clinical stage 4 associated with TB]"
#> 
#> $destinationEntities[[6]]$matchingPVs[[6]]$score
#> [1] -0.5083954
#> 
#> $destinationEntities[[6]]$matchingPVs[[6]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[6]]$matchingPVs[[6]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/1280139157"
#> 
#> $destinationEntities[[6]]$matchingPVs[[6]]$propertyValueType
#> [1] 3
#> 
#> 
#> $destinationEntities[[6]]$matchingPVs[[7]]
#> $destinationEntities[[6]]$matchingPVs[[7]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[6]]$matchingPVs[[7]]$label
#> [1] "HIV- [human immunodeficiency virus] associated dementia complex[HIV disease clinical stage 4 associated with tuberculosis, unspecified]"
#> 
#> $destinationEntities[[6]]$matchingPVs[[7]]$score
#> [1] -0.6303284
#> 
#> $destinationEntities[[6]]$matchingPVs[[7]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[6]]$matchingPVs[[7]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/1280139157"
#> 
#> $destinationEntities[[6]]$matchingPVs[[7]]$propertyValueType
#> [1] 3
#> 
#> 
#> 
#> $destinationEntities[[6]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[6]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[6]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[6]]$chapter
#> [1] "06"
#> 
#> $destinationEntities[[6]]$theCode
#> [1] "6D85.3/1C60.3Z"
#> 
#> $destinationEntities[[6]]$score
#> [1] -0.3301043
#> 
#> $destinationEntities[[6]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[6]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[6]]$entityType
#> [1] 2
#> 
#> $destinationEntities[[6]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[6]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[7]]
#> $destinationEntities[[7]]$id
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1858812010/other"
#> 
#> $destinationEntities[[7]]$title
#> [1] "Other specified HIV disease clinical stage 4 without mention of tuberculosis or malaria"
#> 
#> $destinationEntities[[7]]$stemId
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1858812010/other"
#> 
#> $destinationEntities[[7]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[7]]$postcoordinationAvailability
#> [1] 1
#> 
#> $destinationEntities[[7]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[7]]$hasMaternalChapterLink
#> [1] TRUE
#> 
#> $destinationEntities[[7]]$hasPerinatalChapterLink
#> [1] TRUE
#> 
#> $destinationEntities[[7]]$matchingPVs
#> list()
#> 
#> $destinationEntities[[7]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[7]]$isResidualOther
#> [1] TRUE
#> 
#> $destinationEntities[[7]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[7]]$chapter
#> [1] "01"
#> 
#> $destinationEntities[[7]]$theCode
#> [1] "1C62.3Y"
#> 
#> $destinationEntities[[7]]$score
#> [1] -0.4834757
#> 
#> $destinationEntities[[7]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[7]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[7]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[7]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[7]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[8]]
#> $destinationEntities[[8]]$id
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1173624218 / http://id.who.int/icd/release/11/2024-01/mms/1280139157/other"
#> 
#> $destinationEntities[[8]]$title
#> [1] "Dementia due to human immunodeficiency virus [Other specified HIV disease clinical stage 4 associated with tuberculosis]"
#> 
#> $destinationEntities[[8]]$stemId
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1173624218"
#> 
#> $destinationEntities[[8]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[8]]$postcoordinationAvailability
#> [1] 2
#> 
#> $destinationEntities[[8]]$hasCodingNote
#> [1] TRUE
#> 
#> $destinationEntities[[8]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[8]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[8]]$matchingPVs
#> $destinationEntities[[8]]$matchingPVs[[1]]
#> $destinationEntities[[8]]$matchingPVs[[1]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[8]]$matchingPVs[[1]]$label
#> [1] "HIV - [human immunodeficiency virus] dementia[Other specified HIV disease clinical stage 4 associated with tuberculosis]"
#> 
#> $destinationEntities[[8]]$matchingPVs[[1]]$score
#> [1] -0.5037882
#> 
#> $destinationEntities[[8]]$matchingPVs[[1]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[8]]$matchingPVs[[1]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/1280139157"
#> 
#> $destinationEntities[[8]]$matchingPVs[[1]]$propertyValueType
#> [1] 3
#> 
#> 
#> $destinationEntities[[8]]$matchingPVs[[2]]
#> $destinationEntities[[8]]$matchingPVs[[2]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[8]]$matchingPVs[[2]]$label
#> [1] "HIV- [human immunodeficiency virus] associated cognitive motor complex[Other specified HIV disease clinical stage 4 associated with tuberculosis]"
#> 
#> $destinationEntities[[8]]$matchingPVs[[2]]$score
#> [1] -0.507785
#> 
#> $destinationEntities[[8]]$matchingPVs[[2]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[8]]$matchingPVs[[2]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/1280139157"
#> 
#> $destinationEntities[[8]]$matchingPVs[[2]]$propertyValueType
#> [1] 3
#> 
#> 
#> $destinationEntities[[8]]$matchingPVs[[3]]
#> $destinationEntities[[8]]$matchingPVs[[3]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[8]]$matchingPVs[[3]]$label
#> [1] "HIV- [human immunodeficiency virus] associated dementia complex[Other specified HIV disease clinical stage 4 associated with tuberculosis]"
#> 
#> $destinationEntities[[8]]$matchingPVs[[3]]$score
#> [1] -0.6820793
#> 
#> $destinationEntities[[8]]$matchingPVs[[3]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[8]]$matchingPVs[[3]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/1280139157"
#> 
#> $destinationEntities[[8]]$matchingPVs[[3]]$propertyValueType
#> [1] 3
#> 
#> 
#> 
#> $destinationEntities[[8]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[8]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[8]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[8]]$chapter
#> [1] "06"
#> 
#> $destinationEntities[[8]]$theCode
#> [1] "6D85.3/1C60.3Y"
#> 
#> $destinationEntities[[8]]$score
#> [1] -0.5037882
#> 
#> $destinationEntities[[8]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[8]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[8]]$entityType
#> [1] 2
#> 
#> $destinationEntities[[8]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[8]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[9]]
#> $destinationEntities[[9]]$id
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1173624218 / http://id.who.int/icd/release/11/2024-01/mms/97626683/unspecified"
#> 
#> $destinationEntities[[9]]$title
#> [1] "Dementia due to human immunodeficiency virus [HIV disease clinical stage 4 associated with malaria, unspecified]"
#> 
#> $destinationEntities[[9]]$stemId
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1173624218"
#> 
#> $destinationEntities[[9]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[9]]$postcoordinationAvailability
#> [1] 2
#> 
#> $destinationEntities[[9]]$hasCodingNote
#> [1] TRUE
#> 
#> $destinationEntities[[9]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[9]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[9]]$matchingPVs
#> $destinationEntities[[9]]$matchingPVs[[1]]
#> $destinationEntities[[9]]$matchingPVs[[1]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[9]]$matchingPVs[[1]]$label
#> [1] "HIV - [human immunodeficiency virus] dementia[HIV disease clinical stage 4 associated with malaria, unspecified]"
#> 
#> $destinationEntities[[9]]$matchingPVs[[1]]$score
#> [1] -0.5673337
#> 
#> $destinationEntities[[9]]$matchingPVs[[1]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[9]]$matchingPVs[[1]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/97626683"
#> 
#> $destinationEntities[[9]]$matchingPVs[[1]]$propertyValueType
#> [1] 3
#> 
#> 
#> $destinationEntities[[9]]$matchingPVs[[2]]
#> $destinationEntities[[9]]$matchingPVs[[2]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[9]]$matchingPVs[[2]]$label
#> [1] "HIV- [human immunodeficiency virus] associated cognitive motor complex[HIV disease clinical stage 4 associated with malaria, unspecified]"
#> 
#> $destinationEntities[[9]]$matchingPVs[[2]]$score
#> [1] -0.5713305
#> 
#> $destinationEntities[[9]]$matchingPVs[[2]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[9]]$matchingPVs[[2]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/97626683"
#> 
#> $destinationEntities[[9]]$matchingPVs[[2]]$propertyValueType
#> [1] 3
#> 
#> 
#> $destinationEntities[[9]]$matchingPVs[[3]]
#> $destinationEntities[[9]]$matchingPVs[[3]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[9]]$matchingPVs[[3]]$label
#> [1] "HIV- [human immunodeficiency virus] associated dementia complex[HIV disease clinical stage 4 associated with malaria, unspecified]"
#> 
#> $destinationEntities[[9]]$matchingPVs[[3]]$score
#> [1] -0.7456249
#> 
#> $destinationEntities[[9]]$matchingPVs[[3]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[9]]$matchingPVs[[3]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/97626683"
#> 
#> $destinationEntities[[9]]$matchingPVs[[3]]$propertyValueType
#> [1] 3
#> 
#> 
#> 
#> $destinationEntities[[9]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[9]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[9]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[9]]$chapter
#> [1] "06"
#> 
#> $destinationEntities[[9]]$theCode
#> [1] "6D85.3/1C61.3Z"
#> 
#> $destinationEntities[[9]]$score
#> [1] -0.5673337
#> 
#> $destinationEntities[[9]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[9]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[9]]$entityType
#> [1] 2
#> 
#> $destinationEntities[[9]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[9]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[10]]
#> $destinationEntities[[10]]$id
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1173624218 / http://id.who.int/icd/release/11/2024-01/mms/97626683/other"
#> 
#> $destinationEntities[[10]]$title
#> [1] "Dementia due to human immunodeficiency virus [Other specified HIV disease clinical stage 4 associated with malaria]"
#> 
#> $destinationEntities[[10]]$stemId
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1173624218"
#> 
#> $destinationEntities[[10]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[10]]$postcoordinationAvailability
#> [1] 2
#> 
#> $destinationEntities[[10]]$hasCodingNote
#> [1] TRUE
#> 
#> $destinationEntities[[10]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[10]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[10]]$matchingPVs
#> $destinationEntities[[10]]$matchingPVs[[1]]
#> $destinationEntities[[10]]$matchingPVs[[1]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[10]]$matchingPVs[[1]]$label
#> [1] "HIV - [human immunodeficiency virus] dementia[Other specified HIV disease clinical stage 4 associated with malaria]"
#> 
#> $destinationEntities[[10]]$matchingPVs[[1]]$score
#> [1] -0.6190847
#> 
#> $destinationEntities[[10]]$matchingPVs[[1]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[10]]$matchingPVs[[1]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/97626683"
#> 
#> $destinationEntities[[10]]$matchingPVs[[1]]$propertyValueType
#> [1] 3
#> 
#> 
#> $destinationEntities[[10]]$matchingPVs[[2]]
#> $destinationEntities[[10]]$matchingPVs[[2]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[10]]$matchingPVs[[2]]$label
#> [1] "HIV- [human immunodeficiency virus] associated cognitive motor complex[Other specified HIV disease clinical stage 4 associated with malaria]"
#> 
#> $destinationEntities[[10]]$matchingPVs[[2]]$score
#> [1] -0.6230814
#> 
#> $destinationEntities[[10]]$matchingPVs[[2]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[10]]$matchingPVs[[2]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/97626683"
#> 
#> $destinationEntities[[10]]$matchingPVs[[2]]$propertyValueType
#> [1] 3
#> 
#> 
#> $destinationEntities[[10]]$matchingPVs[[3]]
#> $destinationEntities[[10]]$matchingPVs[[3]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[10]]$matchingPVs[[3]]$label
#> [1] "HIV- [human immunodeficiency virus] associated dementia complex[Other specified HIV disease clinical stage 4 associated with malaria]"
#> 
#> $destinationEntities[[10]]$matchingPVs[[3]]$score
#> [1] -0.7973758
#> 
#> $destinationEntities[[10]]$matchingPVs[[3]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[10]]$matchingPVs[[3]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/97626683"
#> 
#> $destinationEntities[[10]]$matchingPVs[[3]]$propertyValueType
#> [1] 3
#> 
#> 
#> 
#> $destinationEntities[[10]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[10]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[10]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[10]]$chapter
#> [1] "06"
#> 
#> $destinationEntities[[10]]$theCode
#> [1] "6D85.3/1C61.3Y"
#> 
#> $destinationEntities[[10]]$score
#> [1] -0.6190847
#> 
#> $destinationEntities[[10]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[10]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[10]]$entityType
#> [1] 2
#> 
#> $destinationEntities[[10]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[10]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[11]]
#> $destinationEntities[[11]]$id
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1173624218 / http://id.who.int/icd/release/11/2024-01/mms/1858812010/unspecified"
#> 
#> $destinationEntities[[11]]$title
#> [1] "Dementia due to human immunodeficiency virus [HIV disease clinical stage 4 without mention of tuberculosis or malaria, unspecified]"
#> 
#> $destinationEntities[[11]]$stemId
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1173624218"
#> 
#> $destinationEntities[[11]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[11]]$postcoordinationAvailability
#> [1] 2
#> 
#> $destinationEntities[[11]]$hasCodingNote
#> [1] TRUE
#> 
#> $destinationEntities[[11]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[11]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[11]]$matchingPVs
#> $destinationEntities[[11]]$matchingPVs[[1]]
#> $destinationEntities[[11]]$matchingPVs[[1]]$propertyId
#> [1] "Title"
#> 
#> $destinationEntities[[11]]$matchingPVs[[1]]$label
#> [1] "Dementia due to human immunodeficiency virus[HIV disease clinical stage 4 without tuberculosis or malaria]"
#> 
#> $destinationEntities[[11]]$matchingPVs[[1]]$score
#> [1] -0.8556216
#> 
#> $destinationEntities[[11]]$matchingPVs[[1]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[11]]$matchingPVs[[1]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/1858812010"
#> 
#> $destinationEntities[[11]]$matchingPVs[[1]]$propertyValueType
#> [1] 3
#> 
#> 
#> $destinationEntities[[11]]$matchingPVs[[2]]
#> $destinationEntities[[11]]$matchingPVs[[2]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[11]]$matchingPVs[[2]]$label
#> [1] "HIV - [human immunodeficiency virus] dementia[HIV disease clinical stage 4 without tuberculosis or malaria]"
#> 
#> $destinationEntities[[11]]$matchingPVs[[2]]$score
#> [1] -0.8556216
#> 
#> $destinationEntities[[11]]$matchingPVs[[2]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[11]]$matchingPVs[[2]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/1858812010"
#> 
#> $destinationEntities[[11]]$matchingPVs[[2]]$propertyValueType
#> [1] 3
#> 
#> 
#> $destinationEntities[[11]]$matchingPVs[[3]]
#> $destinationEntities[[11]]$matchingPVs[[3]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[11]]$matchingPVs[[3]]$label
#> [1] "HIV- [human immunodeficiency virus] associated cognitive motor complex[HIV disease clinical stage 4 without tuberculosis or malaria]"
#> 
#> $destinationEntities[[11]]$matchingPVs[[3]]$score
#> [1] -0.9194379
#> 
#> $destinationEntities[[11]]$matchingPVs[[3]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[11]]$matchingPVs[[3]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/1858812010"
#> 
#> $destinationEntities[[11]]$matchingPVs[[3]]$propertyValueType
#> [1] 3
#> 
#> 
#> $destinationEntities[[11]]$matchingPVs[[4]]
#> $destinationEntities[[11]]$matchingPVs[[4]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[11]]$matchingPVs[[4]]$label
#> [1] "HIV - [human immunodeficiency virus] dementia[HIV disease clinical stage 4 without mention of tuberculosis or malaria, unspecified]"
#> 
#> $destinationEntities[[11]]$matchingPVs[[4]]$score
#> [1] -0.9675963
#> 
#> $destinationEntities[[11]]$matchingPVs[[4]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[11]]$matchingPVs[[4]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/1858812010"
#> 
#> $destinationEntities[[11]]$matchingPVs[[4]]$propertyValueType
#> [1] 3
#> 
#> 
#> $destinationEntities[[11]]$matchingPVs[[5]]
#> $destinationEntities[[11]]$matchingPVs[[5]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[11]]$matchingPVs[[5]]$label
#> [1] "HIV- [human immunodeficiency virus] associated cognitive motor complex[HIV disease clinical stage 4 without mention of tuberculosis or malaria, unspecified]"
#> 
#> $destinationEntities[[11]]$matchingPVs[[5]]$score
#> [1] -1.031413
#> 
#> $destinationEntities[[11]]$matchingPVs[[5]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[11]]$matchingPVs[[5]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/1858812010"
#> 
#> $destinationEntities[[11]]$matchingPVs[[5]]$propertyValueType
#> [1] 3
#> 
#> 
#> $destinationEntities[[11]]$matchingPVs[[6]]
#> $destinationEntities[[11]]$matchingPVs[[6]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[11]]$matchingPVs[[6]]$label
#> [1] "HIV- [human immunodeficiency virus] associated dementia complex[HIV disease clinical stage 4 without tuberculosis or malaria]"
#> 
#> $destinationEntities[[11]]$matchingPVs[[6]]$score
#> [1] -1.093732
#> 
#> $destinationEntities[[11]]$matchingPVs[[6]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[11]]$matchingPVs[[6]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/1858812010"
#> 
#> $destinationEntities[[11]]$matchingPVs[[6]]$propertyValueType
#> [1] 3
#> 
#> 
#> $destinationEntities[[11]]$matchingPVs[[7]]
#> $destinationEntities[[11]]$matchingPVs[[7]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[11]]$matchingPVs[[7]]$label
#> [1] "HIV- [human immunodeficiency virus] associated dementia complex[HIV disease clinical stage 4 without mention of tuberculosis or malaria, unspecified]"
#> 
#> $destinationEntities[[11]]$matchingPVs[[7]]$score
#> [1] -1.205707
#> 
#> $destinationEntities[[11]]$matchingPVs[[7]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[11]]$matchingPVs[[7]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/1858812010"
#> 
#> $destinationEntities[[11]]$matchingPVs[[7]]$propertyValueType
#> [1] 3
#> 
#> 
#> 
#> $destinationEntities[[11]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[11]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[11]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[11]]$chapter
#> [1] "06"
#> 
#> $destinationEntities[[11]]$theCode
#> [1] "6D85.3/1C62.3Z"
#> 
#> $destinationEntities[[11]]$score
#> [1] -0.8556216
#> 
#> $destinationEntities[[11]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[11]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[11]]$entityType
#> [1] 2
#> 
#> $destinationEntities[[11]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[11]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[12]]
#> $destinationEntities[[12]]$id
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1173624218 / http://id.who.int/icd/release/11/2024-01/mms/1858812010/other"
#> 
#> $destinationEntities[[12]]$title
#> [1] "Dementia due to human immunodeficiency virus [Other specified HIV disease clinical stage 4 without mention of tuberculosis or malaria]"
#> 
#> $destinationEntities[[12]]$stemId
#> [1] "http://id.who.int/icd/release/11/2024-01/mms/1173624218"
#> 
#> $destinationEntities[[12]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[12]]$postcoordinationAvailability
#> [1] 2
#> 
#> $destinationEntities[[12]]$hasCodingNote
#> [1] TRUE
#> 
#> $destinationEntities[[12]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[12]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[12]]$matchingPVs
#> $destinationEntities[[12]]$matchingPVs[[1]]
#> $destinationEntities[[12]]$matchingPVs[[1]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[12]]$matchingPVs[[1]]$label
#> [1] "HIV - [human immunodeficiency virus] dementia[Other specified HIV disease clinical stage 4 without mention of tuberculosis or malaria]"
#> 
#> $destinationEntities[[12]]$matchingPVs[[1]]$score
#> [1] -1.019347
#> 
#> $destinationEntities[[12]]$matchingPVs[[1]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[12]]$matchingPVs[[1]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/1858812010"
#> 
#> $destinationEntities[[12]]$matchingPVs[[1]]$propertyValueType
#> [1] 3
#> 
#> 
#> $destinationEntities[[12]]$matchingPVs[[2]]
#> $destinationEntities[[12]]$matchingPVs[[2]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[12]]$matchingPVs[[2]]$label
#> [1] "HIV- [human immunodeficiency virus] associated cognitive motor complex[Other specified HIV disease clinical stage 4 without mention of tuberculosis or malaria]"
#> 
#> $destinationEntities[[12]]$matchingPVs[[2]]$score
#> [1] -1.083163
#> 
#> $destinationEntities[[12]]$matchingPVs[[2]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[12]]$matchingPVs[[2]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/1858812010"
#> 
#> $destinationEntities[[12]]$matchingPVs[[2]]$propertyValueType
#> [1] 3
#> 
#> 
#> $destinationEntities[[12]]$matchingPVs[[3]]
#> $destinationEntities[[12]]$matchingPVs[[3]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[12]]$matchingPVs[[3]]$label
#> [1] "HIV- [human immunodeficiency virus] associated dementia complex[Other specified HIV disease clinical stage 4 without mention of tuberculosis or malaria]"
#> 
#> $destinationEntities[[12]]$matchingPVs[[3]]$score
#> [1] -1.257458
#> 
#> $destinationEntities[[12]]$matchingPVs[[3]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[12]]$matchingPVs[[3]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1173624218 / http://id.who.int/icd/entity/1858812010"
#> 
#> $destinationEntities[[12]]$matchingPVs[[3]]$propertyValueType
#> [1] 3
#> 
#> 
#> 
#> $destinationEntities[[12]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[12]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[12]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[12]]$chapter
#> [1] "06"
#> 
#> $destinationEntities[[12]]$theCode
#> [1] "6D85.3/1C62.3Y"
#> 
#> $destinationEntities[[12]]$score
#> [1] -1.019347
#> 
#> $destinationEntities[[12]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[12]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[12]]$entityType
#> [1] 2
#> 
#> $destinationEntities[[12]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[12]]$descendants
#> list()
#> 
#> 
#> 

# Assuming that the CLIENT ID and CLIENT SECRET are set up. Substitute accordingly
if (exists("CLIENT_ID") & exists("CLIENT_SECRET")) {
  # Generated token
  token <- get_token(CLIENT_ID, CLIENT_SECRET)

  # Search for ICD-10 releases
  icd11_search2(token, text = "Cerebrovascular accident")
}
#> $error
#> [1] FALSE
#> 
#> $errorMessage
#> NULL
#> 
#> $resultChopped
#> [1] FALSE
#> 
#> $wordSuggestionsChopped
#> [1] FALSE
#> 
#> $guessType
#> [1] 2
#> 
#> $uniqueSearchId
#> [1] "ffe6d1ca-bdd1-49c4-a943-84a27e6d7051"
#> 
#> $words
#> NULL
#> 
#> $destinationEntities
#> $destinationEntities[[1]]
#> $destinationEntities[[1]]$id
#> [1] "http://id.who.int/icd/entity/1442995018"
#> 
#> $destinationEntities[[1]]$title
#> [1] "Stroke not known if ischaemic or haemorrhagic"
#> 
#> $destinationEntities[[1]]$stemId
#> [1] "http://id.who.int/icd/entity/1442995018"
#> 
#> $destinationEntities[[1]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[1]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[1]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$matchingPVs
#> $destinationEntities[[1]]$matchingPVs[[1]]
#> $destinationEntities[[1]]$matchingPVs[[1]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[1]]$matchingPVs[[1]]$label
#> [1] "cerebrovascular accident"
#> 
#> $destinationEntities[[1]]$matchingPVs[[1]]$score
#> [1] 1
#> 
#> $destinationEntities[[1]]$matchingPVs[[1]]$important
#> [1] TRUE
#> 
#> $destinationEntities[[1]]$matchingPVs[[1]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1442995018"
#> 
#> $destinationEntities[[1]]$matchingPVs[[1]]$propertyValueType
#> [1] 0
#> 
#> 
#> $destinationEntities[[1]]$matchingPVs[[2]]
#> $destinationEntities[[1]]$matchingPVs[[2]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[1]]$matchingPVs[[2]]$label
#> [1] "acute cerebrovascular accident"
#> 
#> $destinationEntities[[1]]$matchingPVs[[2]]$score
#> [1] 0.7785942
#> 
#> $destinationEntities[[1]]$matchingPVs[[2]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$matchingPVs[[2]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1442995018"
#> 
#> $destinationEntities[[1]]$matchingPVs[[2]]$propertyValueType
#> [1] 0
#> 
#> 
#> $destinationEntities[[1]]$matchingPVs[[3]]
#> $destinationEntities[[1]]$matchingPVs[[3]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[1]]$matchingPVs[[3]]$label
#> [1] "massive cerebrovascular accident"
#> 
#> $destinationEntities[[1]]$matchingPVs[[3]]$score
#> [1] 0.5094063
#> 
#> $destinationEntities[[1]]$matchingPVs[[3]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$matchingPVs[[3]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1442995018"
#> 
#> $destinationEntities[[1]]$matchingPVs[[3]]$propertyValueType
#> [1] 0
#> 
#> 
#> $destinationEntities[[1]]$matchingPVs[[4]]
#> $destinationEntities[[1]]$matchingPVs[[4]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[1]]$matchingPVs[[4]]$label
#> [1] "CVA - [cerebrovascular accident]"
#> 
#> $destinationEntities[[1]]$matchingPVs[[4]]$score
#> [1] 0.3315121
#> 
#> $destinationEntities[[1]]$matchingPVs[[4]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$matchingPVs[[4]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1442995018"
#> 
#> $destinationEntities[[1]]$matchingPVs[[4]]$propertyValueType
#> [1] 0
#> 
#> 
#> 
#> $destinationEntities[[1]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$chapter
#> [1] "08"
#> 
#> $destinationEntities[[1]]$theCode
#> NULL
#> 
#> $destinationEntities[[1]]$score
#> [1] 1
#> 
#> $destinationEntities[[1]]$titleIsASearchResult
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$titleIsTopScore
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[1]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[1]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[2]]
#> $destinationEntities[[2]]$id
#> [1] "http://id.who.int/icd/entity/1614716959"
#> 
#> $destinationEntities[[2]]$title
#> [1] "Cerebral ischaemic stroke of unknown cause"
#> 
#> $destinationEntities[[2]]$stemId
#> [1] "http://id.who.int/icd/entity/1614716959"
#> 
#> $destinationEntities[[2]]$isLeaf
#> [1] FALSE
#> 
#> $destinationEntities[[2]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[2]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[2]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[2]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[2]]$matchingPVs
#> $destinationEntities[[2]]$matchingPVs[[1]]
#> $destinationEntities[[2]]$matchingPVs[[1]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[2]]$matchingPVs[[1]]$label
#> [1] "ischaemic cerebrovascular accident"
#> 
#> $destinationEntities[[2]]$matchingPVs[[1]]$score
#> [1] 0.6584358
#> 
#> $destinationEntities[[2]]$matchingPVs[[1]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[2]]$matchingPVs[[1]]$foundationUri
#> [1] "http://id.who.int/icd/entity/1614716959"
#> 
#> $destinationEntities[[2]]$matchingPVs[[1]]$propertyValueType
#> [1] 0
#> 
#> 
#> 
#> $destinationEntities[[2]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[2]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[2]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[2]]$chapter
#> [1] "08"
#> 
#> $destinationEntities[[2]]$theCode
#> NULL
#> 
#> $destinationEntities[[2]]$score
#> [1] 0.6584358
#> 
#> $destinationEntities[[2]]$titleIsASearchResult
#> [1] FALSE
#> 
#> $destinationEntities[[2]]$titleIsTopScore
#> [1] FALSE
#> 
#> $destinationEntities[[2]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[2]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[2]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[3]]
#> $destinationEntities[[3]]$id
#> [1] "http://id.who.int/icd/entity/557175275"
#> 
#> $destinationEntities[[3]]$title
#> [1] "Late effects of stroke not known if ischaemic or haemorrhagic"
#> 
#> $destinationEntities[[3]]$stemId
#> [1] "http://id.who.int/icd/entity/557175275"
#> 
#> $destinationEntities[[3]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[3]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[3]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$matchingPVs
#> $destinationEntities[[3]]$matchingPVs[[1]]
#> $destinationEntities[[3]]$matchingPVs[[1]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[3]]$matchingPVs[[1]]$label
#> [1] "old cerebrovascular accident"
#> 
#> $destinationEntities[[3]]$matchingPVs[[1]]$score
#> [1] 0.5563917
#> 
#> $destinationEntities[[3]]$matchingPVs[[1]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$matchingPVs[[1]]$foundationUri
#> [1] "http://id.who.int/icd/entity/557175275"
#> 
#> $destinationEntities[[3]]$matchingPVs[[1]]$propertyValueType
#> [1] 0
#> 
#> 
#> $destinationEntities[[3]]$matchingPVs[[2]]
#> $destinationEntities[[3]]$matchingPVs[[2]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[3]]$matchingPVs[[2]]$label
#> [1] "past cerebrovascular accident"
#> 
#> $destinationEntities[[3]]$matchingPVs[[2]]$score
#> [1] 0.5527899
#> 
#> $destinationEntities[[3]]$matchingPVs[[2]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$matchingPVs[[2]]$foundationUri
#> [1] "http://id.who.int/icd/entity/557175275"
#> 
#> $destinationEntities[[3]]$matchingPVs[[2]]$propertyValueType
#> [1] 0
#> 
#> 
#> $destinationEntities[[3]]$matchingPVs[[3]]
#> $destinationEntities[[3]]$matchingPVs[[3]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[3]]$matchingPVs[[3]]$label
#> [1] "previous cerebrovascular accident"
#> 
#> $destinationEntities[[3]]$matchingPVs[[3]]$score
#> [1] 0.5289606
#> 
#> $destinationEntities[[3]]$matchingPVs[[3]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$matchingPVs[[3]]$foundationUri
#> [1] "http://id.who.int/icd/entity/557175275"
#> 
#> $destinationEntities[[3]]$matchingPVs[[3]]$propertyValueType
#> [1] 0
#> 
#> 
#> $destinationEntities[[3]]$matchingPVs[[4]]
#> $destinationEntities[[3]]$matchingPVs[[4]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[3]]$matchingPVs[[4]]$label
#> [1] "old CVA - [cerebrovascular accident]"
#> 
#> $destinationEntities[[3]]$matchingPVs[[4]]$score
#> [1] -0.1120962
#> 
#> $destinationEntities[[3]]$matchingPVs[[4]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$matchingPVs[[4]]$foundationUri
#> [1] "http://id.who.int/icd/entity/557175275"
#> 
#> $destinationEntities[[3]]$matchingPVs[[4]]$propertyValueType
#> [1] 0
#> 
#> 
#> 
#> $destinationEntities[[3]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$chapter
#> [1] "08"
#> 
#> $destinationEntities[[3]]$theCode
#> NULL
#> 
#> $destinationEntities[[3]]$score
#> [1] 0.5563917
#> 
#> $destinationEntities[[3]]$titleIsASearchResult
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$titleIsTopScore
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[3]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[3]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[4]]
#> $destinationEntities[[4]]$id
#> [1] "http://id.who.int/icd/entity/873092535"
#> 
#> $destinationEntities[[4]]$title
#> [1] "Intracerebral haemorrhage"
#> 
#> $destinationEntities[[4]]$stemId
#> [1] "http://id.who.int/icd/entity/873092535"
#> 
#> $destinationEntities[[4]]$isLeaf
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[4]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$matchingPVs
#> $destinationEntities[[4]]$matchingPVs[[1]]
#> $destinationEntities[[4]]$matchingPVs[[1]]$propertyId
#> [1] "Synonym"
#> 
#> $destinationEntities[[4]]$matchingPVs[[1]]$label
#> [1] "haemorrhagic cerebrovascular accident"
#> 
#> $destinationEntities[[4]]$matchingPVs[[1]]$score
#> [1] 0.1609707
#> 
#> $destinationEntities[[4]]$matchingPVs[[1]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$matchingPVs[[1]]$foundationUri
#> [1] "http://id.who.int/icd/entity/873092535"
#> 
#> $destinationEntities[[4]]$matchingPVs[[1]]$propertyValueType
#> [1] 0
#> 
#> 
#> 
#> $destinationEntities[[4]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$chapter
#> [1] "08"
#> 
#> $destinationEntities[[4]]$theCode
#> NULL
#> 
#> $destinationEntities[[4]]$score
#> [1] 0.1609707
#> 
#> $destinationEntities[[4]]$titleIsASearchResult
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$titleIsTopScore
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[4]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[4]]$descendants
#> list()
#> 
#> 
#> $destinationEntities[[5]]
#> $destinationEntities[[5]]$id
#> [1] "http://id.who.int/icd/entity/209094863"
#> 
#> $destinationEntities[[5]]$title
#> [1] "Observation for suspected cerebrovascular accident, ruled out"
#> 
#> $destinationEntities[[5]]$stemId
#> [1] "http://id.who.int/icd/entity/209094863"
#> 
#> $destinationEntities[[5]]$isLeaf
#> [1] TRUE
#> 
#> $destinationEntities[[5]]$postcoordinationAvailability
#> [1] 0
#> 
#> $destinationEntities[[5]]$hasCodingNote
#> [1] FALSE
#> 
#> $destinationEntities[[5]]$hasMaternalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[5]]$hasPerinatalChapterLink
#> [1] FALSE
#> 
#> $destinationEntities[[5]]$matchingPVs
#> list()
#> 
#> $destinationEntities[[5]]$propertiesTruncated
#> [1] FALSE
#> 
#> $destinationEntities[[5]]$isResidualOther
#> [1] FALSE
#> 
#> $destinationEntities[[5]]$isResidualUnspecified
#> [1] FALSE
#> 
#> $destinationEntities[[5]]$chapter
#> [1] "24"
#> 
#> $destinationEntities[[5]]$theCode
#> NULL
#> 
#> $destinationEntities[[5]]$score
#> [1] -1.462644
#> 
#> $destinationEntities[[5]]$titleIsASearchResult
#> [1] TRUE
#> 
#> $destinationEntities[[5]]$titleIsTopScore
#> [1] TRUE
#> 
#> $destinationEntities[[5]]$entityType
#> [1] 0
#> 
#> $destinationEntities[[5]]$important
#> [1] FALSE
#> 
#> $destinationEntities[[5]]$descendants
#> list()
#> 
#> 
#>